% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment_networks.R
\name{add_line_breaks}
\alias{add_line_breaks}
\title{Add line breaks to a string}
\usage{
add_line_breaks(x, max_char = 50, sep = ",")
}
\arguments{
\item{x}{character string}

\item{max_char}{integer, add line breaks at least every \code{max_char} characters}

\item{sep}{character after which to add line break}
}
\value{
string with additional \<br\> if necessary
}
\description{
Given a long string, add line breaks "\<br\>" at specified intervals.
Breaks are always added at the preceding instance of the separator.
Function is used internally in \code{\link[=enrichment_network_vis]{enrichment_network_vis()}}.
}
\seealso{
\code{\link[=enrichment_network_vis]{enrichment_network_vis()}}
}
\keyword{internal}
