% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_graphical_clustering.R
\name{annotate_cluster_res}
\alias{annotate_cluster_res}
\title{Annotate clustering results}
\usage{
annotate_cluster_res(cluster_res)
}
\arguments{
\item{cluster_res}{data frame with at least two columns: "feature" and "cluster".
The "feature" column must be in the format
"\link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV};\link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV};feature_ID".}
}
\value{
the input as a data frame with additional "ome", "tissue", and "feature_ID" columns
and unfactorized "cluster" and "feature" columns
}
\description{
Add "ome", "tissue", and "feature_ID" columns to the input.
Run within \code{\link[=check_cluster_res_format]{check_cluster_res_format()}}.
}
\seealso{
\code{\link[=check_cluster_res_format]{check_cluster_res_format()}}
}
