% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_params.R
\name{assay}
\alias{assay}
\title{Internal functions used to define parameters and column descriptions.
Please add functions in alphabetical order.}
\usage{
assay()
}
\description{
To render:
tissue character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}
assay character, assay abbreviation, one of \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV}
Type:
tissue @eval tissue()
assay @eval assay()
}
