% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_outliers.R
\name{atac_call_outliers}
\alias{atac_call_outliers}
\title{Call ATAC-seq sample outliers}
\usage{
atac_call_outliers(tissues, scratchdir = ".")
}
\arguments{
\item{tissues}{character vector of tissue abbreviations for which to call ATAC-seq outliers,
one of "BAT", "HEART", "HIPPOC", "KIDNEY", "LIVER", "LUNG", "SKM-GN", "WAT-SC"}

\item{scratchdir}{character, local directory in which to download data from Google Cloud Storage.
Current working directory by default.}
}
\value{
NULL if there are no outliers, or a data frame with three columns and one row per outlier:
\describe{
\item{\code{viallabel}}{character, sample identifier}
\item{\code{tissue}}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}
\item{\code{reason}}{character, PC(s) in which the sample was flagged}
}
}
\description{
Identify samples that are outside of 3 times the interquartile range of principal
components that explain at least 7.5\% of variance in each tissue. Use only the 10,000 most variable peaks.
Outlier calling is performed separately in each sex.
This specifies ATAC-seq outliers excluded from differential analysis by MoTrPAC.
}
\examples{
\dontrun{
atac_call_outliers("LIVER","/tmp")
atac_call_outliers(c("LIVER","BAT"),"/tmp")
}
}
\seealso{
\code{\link[=call_pca_outliers]{call_pca_outliers()}} for workhorse function and \code{\link[=plot_pcs]{plot_pcs()}} for plotting function
}
