% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_prep_data.R
\name{atac_prep_data}
\alias{atac_prep_data}
\title{Prepare ATAC-seq dataset}
\usage{
atac_prep_data(
  tissue,
  sex = NULL,
  covariates = c("Sample_batch", "peak_enrich.frac_reads_in_peaks.macs2.frip"),
  filter_counts = FALSE,
  return_normalized_data = FALSE,
  scratchdir = ".",
  outliers = data.table::data.table(MotrpacRatTraining6moData::OUTLIERS)[assay == "ATAC",
    viallabel],
  nrows = Inf
)
}
\arguments{
\item{tissue}{character, tissue abbreviation, one of "BAT", "HEART", "HIPPOC", "KIDNEY", "LIVER", "LUNG", "SKM-GN", "WAT-SC"}

\item{sex}{character, one of 'male' or 'female'}

\item{covariates}{character vector of covariates that correspond to column names of \link[MotrpacRatTraining6moData:ATAC_META]{MotrpacRatTraining6moData::ATAC_META}.
Defaults to covariates that were used for the manuscript.}

\item{filter_counts}{bool, whether to return filtered raw counts}

\item{return_normalized_data}{bool, whether to also return normalized data}

\item{scratchdir}{character, local directory in which to download data from
Google Cloud Storage. Current working directory by default.}

\item{outliers}{vector of viallabels to exclude from the returned data. Defaults
to \code{[MotrpacRatTraining6moData::OUTLIERS]$viallabel}}

\item{nrows}{integer, number of rows to return. Defaults to Inf. Useful to return a subset of a large data frame for tests.}
}
\value{
named list of five items:
\describe{
\item{\code{metadata}}{data frame of combined \link[MotrpacRatTraining6moData:PHENO]{MotrpacRatTraining6moData::PHENO} and
\link[MotrpacRatTraining6moData:ATAC_META]{MotrpacRatTraining6moData::ATAC_META}, filtered to samples in \code{tissue}.}
\item{\code{covariates}}{character vector of covariates to adjust for during differential analysis, same as input}
\item{\code{raw_counts}}{data frame of raw counts with feature IDs
as row names and vial labels as column names. See \link[MotrpacRatTraining6moData:ATAC_RAW_COUNTS]{MotrpacRatTraining6moData::ATAC_RAW_COUNTS} for details.}
\item{\code{norm_data}}{data frame of quantile-normalized data with feature IDs as row names and vial labels as column names.
See \link[MotrpacRatTraining6moData:ATAC_NORM_DATA]{MotrpacRatTraining6moData::ATAC_NORM_DATA} for details.}
\item{\code{outliers}}{subset of \code{outliers} in input removed from the data}
}
}
\description{
Retrieve and format ATAC-seq sample-level data and metadata for a given tissue.
}
\examples{
\dontrun{
# Process gastrocnemius ATAC-seq data with default parameters, i.e., return data from both 
# sexes, remove established outliers, download data to current working directory
gastroc_data1 = atac_prep_data("SKM-GN")

# Same as above but do not remove outliers if they exist 
gastroc_data2 = atac_prep_data("SKM-GN", outliers = NULL)

# Same as above but only return data from male samples
gastroc_data3 = atac_prep_data("SKM-GN", outliers = NULL, sex = "male")
}
}
