% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atac_differential_analysis.R
\name{atac_timewise_da}
\alias{atac_timewise_da}
\title{ATAC-seq timewise differential analysis}
\usage{
atac_timewise_da(
  tissue,
  covariates = c("Sample_batch", "peak_enrich.frac_reads_in_peaks.macs2.frip"),
  outliers =
    na.omit(MotrpacRatTraining6moData::OUTLIERS$viallabel[MotrpacRatTraining6moData::OUTLIERS$assay
    == "ATAC"]),
  scratchdir = ".",
  rdata_outfile = NULL,
  overwrite = FALSE,
  verbose = TRUE,
  n_features = Inf
)
}
\arguments{
\item{tissue}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}

\item{covariates}{character vector of covariates that correspond to column names of \link[MotrpacRatTraining6moData:ATAC_META]{MotrpacRatTraining6moData::ATAC_META}.
Defaults to covariates that were used for the manuscript.}

\item{outliers}{vector of viallabels to exclude during differential analysis. Defaults
to \code{\link[MotrpacRatTraining6moData]{OUTLIERS}$viallabel}}

\item{scratchdir}{character, local directory in which to download data from
Google Cloud Storage. Current working directory by default.}

\item{rdata_outfile}{NULL or path in which to save eBayes objects in an RData file}

\item{overwrite}{boolean, whether to overwrite the file if \code{rdata_outfile} exists}

\item{verbose}{boolean, whether to print messages}

\item{n_features}{integer, number of peaks to perform differential analysis on. Useful for troubleshooting.}
}
\value{
a data frame with one row per feature:
\describe{
\item{\code{feature_ID}}{character, MoTrPAC feature identifier}
\item{\code{sex}}{character, one of 'male' or 'female'}
\item{\code{comparison_group}}{character, time point of trained animals compared to the sex-matched sedentary control animals, one of '1w', '2w', '4w', '8w'}
\item{\code{assay}}{character, assay abbreviation, one of \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV}}
\item{\code{assay_code}}{character, assay code used in data release. See \link[MotrpacBicQC:assay_codes]{MotrpacBicQC::assay_codes}.}
\item{\code{tissue}}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}
\item{\code{tissue_code}}{character, tissue code used in data release. See \link[MotrpacBicQC:bic_animal_tissue_code]{MotrpacBicQC::bic_animal_tissue_code}.}
\item{\code{covariates}}{character, comma-separated list of adjustment variables}
\item{\code{removed_samples}}{character, comma-separated list of outliers (vial labels) removed from differential analysis}
\item{\code{logFC}}{double, log fold-change where the numerator is 'comparison_group', e.g., '1w', and the denominator is the group of sex-matched sedentary control animals}
\item{\code{logFC_se}}{double, standard error of the log fold-change}
\item{\code{tscore}}{double, t statistic}
\item{\code{p_value}}{double, unadjusted p-value for the difference between 'comparison_group' and the group of sex-matched sedentary control animals}
}
}
\description{
Use limma to perform pairwise contrasts between each group of trained animals
and the sex-matched control group for a single tissue. Analysis is performed separately for males and
females.
}
\examples{
\dontrun{
# Perform timewise differential analysis for chromatin accessibility peaks measured in hippocampus with default parameters.
hippoc_atac_da = atac_timewise_da("HIPPOC")
}
}
