% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atac_differential_analysis.R
\name{atac_training_da}
\alias{atac_training_da}
\title{ATAC-seq training differential analysis}
\usage{
atac_training_da(
  tissue,
  covariates = c("Sample_batch", "peak_enrich.frac_reads_in_peaks.macs2.frip"),
  outliers = data.table::data.table(MotrpacRatTraining6moData::OUTLIERS)[assay == "ATAC",
    viallabel],
  scratchdir = ".",
  rdata_outfile = NULL,
  overwrite = FALSE,
  verbose = TRUE,
  n_features = Inf
)
}
\arguments{
\item{tissue}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}

\item{covariates}{character vector of covariates that correspond to column names of \link[MotrpacRatTraining6moData:ATAC_META]{MotrpacRatTraining6moData::ATAC_META}.
Defaults to covariates that were used for the manuscript.}

\item{outliers}{vector of viallabels to exclude during differential analysis. Defaults
to \code{[MotrpacRatTraining6moData::OUTLIERS]$viallabel}}

\item{scratchdir}{character, local directory in which to download data from
Google Cloud Storage. Current working directory by default.}

\item{rdata_outfile}{NULL or path in which to save eBayes objects in an RData file}

\item{overwrite}{boolean, whether to overwrite the file if \code{rdata_outfile} exists}

\item{verbose}{boolean, whether to print messages}

\item{n_features}{integer, number of peaks to perform differential analysis on. Useful for troubleshooting.}
}
\value{
data frame with one row per feature:
\describe{
\item{\code{feature_ID}}{character, MoTrPAC feature identifier}
\item{\code{assay}}{character, assay abbreviation, one of \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV}}
\item{\code{assay_code}}{character, assay code used in data release. See \link[MotrpacBicQC:assay_codes]{MotrpacBicQC::assay_codes}.}
\item{\code{tissue}}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}
\item{\code{tissue_code}}{character, tissue code used in data release. See \link[MotrpacBicQC:bic_animal_tissue_code]{MotrpacBicQC::bic_animal_tissue_code}.}
\item{\code{removed_samples_male}}{character, comma-separated list of male outliers (vial labels) removed from differential analysis}
\item{\code{removed_samples_female}}{character, comma-separated list of female outliers (vial labels) removed from differential analysis}
\item{\code{fscore_male}}{double, F statistic for males}
\item{\code{fscore_female}}{double, F statistic for females}
\item{\code{p_value_male}}{double, nominal F-test p-value for males}
\item{\code{p_value_female}}{double, nominal F-test p-value for females}
\item{\code{full_model_male}}{character, full model used in F-test for males}
\item{\code{full_model_female}}{character, full model used in F-test for females}
\item{\code{reduced_model_male}}{character, reduced model used in F-test for males}
\item{\code{reduced_model_female}}{character, reduced model used in F-test for females}
\item{\code{p_value}}{double, combined male and female nominal p-value using the sum of logs}
}
}
\description{
Use limma to perform an F-test to test the effect of training
across time points. Analysis is performed separately for males and females.
}
\examples{

\dontrun{
# Perform differential analysis for chromatin accessibility peaks in 
# brown adipose tissue with default parameters, 
# i.e., outliers and covariates used for the manuscript
da = atac_training_da("BAT")

# Same as above but save the [limma::eBayes] MArrayLM objects in an RData file 
da = atac_training_da("BAT", rdata_outfile = "~/test/BAT_ATAC_training-da.RData", overwrite = TRUE)
}
}
