% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment_networks.R
\name{calc_similarity_metric}
\alias{calc_similarity_metric}
\title{Calculate pathway similarity metric}
\usage{
calc_similarity_metric(string1, string2)
}
\arguments{
\item{string1}{character, comma-separated list of intersection with pathway1}

\item{string2}{character, comma-separated list of intersection with pathway2}
}
\value{
numeric similarity score
}
\description{
Calculate EnrichmentMap's similarity metric for pathways, which is 50\% Jaccard
index and 50\% overlap score.
Function is used internally in \code{\link[=enrichment_network_vis]{enrichment_network_vis()}}.
}
\seealso{
\code{\link[=enrichment_network_vis]{enrichment_network_vis()}}
}
