% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_outliers.R
\name{call_pca_outliers}
\alias{call_pca_outliers}
\title{Use PCA method to call outliers}
\usage{
call_pca_outliers(
  norm,
  min_pc_ve,
  plot,
  verbose,
  iqr_coef = 3,
  N = Inf,
  TITLE = NULL
)
}
\arguments{
\item{norm}{filtered, normalized values}

\item{min_pc_ve}{minimum percent variance explained by a PC to check it for outliers}

\item{plot}{bool, whether or not to print plots}

\item{verbose}{bool, whether or not to print descriptive strings}

\item{iqr_coef}{numeric, flag PC outliers if they are outside of IQR * \code{iqr_coef}}

\item{N}{integer, select N most variable features}

\item{TITLE}{string, substring to include in PC plot titles}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO

}
