% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_outliers.R
\name{call_transcript_outliers}
\alias{call_transcript_outliers}
\title{Call RNA-seq outliers}
\usage{
call_transcript_outliers(tissue_codes)
}
\arguments{
\item{tissues}{list of tissue abbreviations for which to call RNA-seq outliers. 
See [tissue_abbreviations()] for accepted values.}
}
\value{
something
}
\description{
Description
}
\examples{
call_transcript_outliers("t55-gastrocnemius")
call_transcript_outliers(c("t55-gastrocnemius","t56-vastus-lateralis"))

}
\seealso{
[call_pca_outliers()] for workhorse function, [plot_pcs()] for plotting function, 
    [tissue_abbreviations()] for list of accepted tissue abbrevations
}
