% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_graphical_clustering.R
\name{check_cluster_res_format}
\alias{check_cluster_res_format}
\title{Check clustering results format}
\usage{
check_cluster_res_format(cluster_res)
}
\arguments{
\item{cluster_res}{Either a data frame or a list of lists.
If a data frame, it needs at least two columns: "feature" and "cluster".
The "feature" column should be in the format
'\link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV};\link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV};feature_ID'.
If a list of lists, each sublist must be named with the cluster name (character string),
and the values must be features in the format
'\link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV};\link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV};feature_ID'.}
}
\value{
a data frame with 5 columns:
\describe{
\item{\code{feature}}{character, unique feature identifier in the format '\link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV};\link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV};feature_ID' only for training-regulated features at 5\% IHW FDR. For redundant differential features, 'feature_ID' is prepended with the specific platform to make unique identifiers. See \link[MotrpacRatTraining6moData:REPEATED_FEATURES]{MotrpacRatTraining6moData::REPEATED_FEATURES} for details.}
\item{\code{cluster}}{character, cluster label}
\item{\code{ome}}{character, assay abbreviation, one of \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV}}
\item{\code{tissue}}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}
\item{\code{feature_ID}}{character, MoTrPAC feature identifier}
}
}
\description{
This function will generate warnings or errors if there are issues with the format of the input.
Used internally in \code{\link[=extract_main_clusters]{extract_main_clusters()}}.
}
\examples{
# Note this example is redundant because check_cluster_res_format()
# is called within extract_main_clusters(), but it provides an 
# example input and output
cluster_res = extract_main_clusters()
cluster_res_checked = check_cluster_res_format(cluster_res)

}
\seealso{
\code{\link[=extract_main_clusters]{extract_main_clusters()}}
}
