% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_da_args}
\alias{check_da_args}
\title{Check arguments for DA functions}
\usage{
check_da_args(tissue, outfile, overwrite, outfile_is_rdata = TRUE)
}
\arguments{
\item{tissue}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}

\item{outfile}{character, output file}

\item{overwrite}{bool, whether to overwrite \code{outfile} if it exists}

\item{outfile_is_rdata}{bool, whether \code{outfile} is intended to save RData}
}
\description{
Internal function used to check arguments for differential analysis functions
}
