% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_dea_args}
\alias{check_dea_args}
\title{TODO
Internal function used to check arguments for differential analysis functions}
\usage{
check_dea_args(tissue, outfile, overwrite, outfile_is_rdata = TRUE)
}
\arguments{
\item{tissue}{tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}

\item{outfile}{}

\item{overwrite}{}

\item{outfile_is_rdata}{}
}
\description{
TODO
Internal function used to check arguments for differential analysis functions
}
\examples{
TODO 

}
