% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment_networks.R
\name{collapse_p}
\alias{collapse_p}
\title{Collapse p-values}
\usage{
collapse_p(ps)
}
\arguments{
\item{ps}{vector of nominal p-values}
}
\value{
merged p-value
}
\description{
Collapse p-values using \code{\link[metap:sumlog]{metap::sumlog()}} when multiple are present.
Function is used internally in \code{\link[=enrichment_network_vis]{enrichment_network_vis()}}.
}
\seealso{
\code{\link[=enrichment_network_vis]{enrichment_network_vis()}}
}
