% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_data.R
\name{combine_da_results}
\alias{combine_da_results}
\title{Combine differential analysis results}
\usage{
combine_da_results()
}
\arguments{
\item{tissues}{optional character vector of tissue abbreviations, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}.}

\item{assays}{optional character vector of assay abbreviations, one of \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV}}

\item{training_regulated_only}{bool, whether to return data only for training-regulated features}
}
\value{
data frame... TODO
}
\description{
Combine differential analysis results from the specified tissues and omes(s)/assay(s).
If no tissues or omes are specified,
all differential analysis results are returned.
}
\details{
For \strong{all} epigenetic differential analysis results, see the
documentation for \link[MotrpacRatTraining6moData:ATAC_DA]{MotrpacRatTraining6moData::ATAC_DA} and
\link[MotrpacRatTraining6moData:METHYL_DA]{MotrpacRatTraining6moData::METHYL_DA}, e.g., \code{?ATAC_DA}.
}
