% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_prep_data.R
\name{combine_normalized_data}
\alias{combine_normalized_data}
\title{Combine normalized sample-level data}
\usage{
combine_normalized_data(
  tissues = MotrpacRatTraining6moData::TISSUE_ABBREV,
  assays = MotrpacRatTraining6moData::ASSAY_ABBREV,
  include_epigen = FALSE,
  scratchdir = ".",
  training_regulated_only = FALSE,
  exclude_outliers = FALSE,
  nrows = Inf
)
}
\arguments{
\item{tissues}{optional character vector of tissue abbreviations,
one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}.}

\item{assays}{optional character vector of assay abbreviations,
one of \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV}}

\item{include_epigen}{bool, whether to include the full ATAC or METHYL
differential analysis results from Google Cloud Storage.
Only relevant if \code{assays} includes "ATAC" or "METHYL".
\code{FALSE} by default.}

\item{scratchdir}{character, local directory in which to download data from the web.
Current working directory by default. Only relevant if \code{assays} includes "ATAC" or "METHYL".}

\item{training_regulated_only}{bool, whether to filter features down to those training-regulated at 5\% FDR}

\item{exclude_outliers}{bool, whether to remove sample outliers specified by \link[MotrpacRatTraining6moData:OUTLIERS]{MotrpacRatTraining6moData::OUTLIERS}}

\item{nrows}{integer, number of rows to return from each dataset. Defaults to Inf.
Useful to return a subset of a large data frame for tests.}
}
\value{
data frame with features in rows and Participant IDs (PIDs) in columns
}
\description{
Combine normalized sample-level data from the specified tissues and omes(s)/assay(s).
If no tissues or omes are specified, all data is returned.
In order to combine data from different tissues and data types, sample-specific vial labels
are converted to animal-specific Participant IDs (PIDs).
}
\examples{
# Return all normalized RNA-seq data
data = combine_normalized_data(assays = "TRNSCRPT")

# Return all normalized proteomics data. Exclude outliers 
data = combine_normalized_data(assays = c("PROT","UBIQ","PHOSPHO","ACETYL"),
                               exclude_outliers = TRUE)

# Return normalized ATAC-seq data for training-regulated features 
data = combine_normalized_data(assays = "ATAC", training_regulated_only = TRUE)

# Return normalized ATAC-seq data for the first 1000 features in each tissue 
data = combine_normalized_data(assays = "ATAC", 
                               nrows = 1000, 
                               scratchdir = "/tmp", 
                               include_epigen = TRUE)

# Return all normalized metabolomics data 
data = combine_normalized_data(assays = "METAB")
}
\seealso{
\code{\link[=load_sample_data]{load_sample_data()}}, \code{\link[=viallabel_to_pid]{viallabel_to_pid()}}
}
