% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_data.R
\name{combine_sample_level_data}
\alias{combine_sample_level_data}
\title{Combine sample-level data}
\usage{
combine_sample_level_data(
  tissues = NULL,
  assays = NULL,
  training_regulated_only = TRUE
)
}
\arguments{
\item{tissues}{optional character vector of tissue abbreviations, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}.}

\item{assays}{optional character vector of assay abbreviations, one of \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV}}

\item{training_regulated_only}{bool, whether to return data only for training-regulated features}
}
\value{
data frame... TODO
}
\description{
Combine data from the specified tissues and omes(s)/assay(s).
If no tissues or omes are specified, all data is returned.
}
\examples{
print("TODO")
}
