% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment.R
\name{custom_cluster_pathway_enrichment}
\alias{custom_cluster_pathway_enrichment}
\title{Custom pathway enrichment for graphical clusters}
\usage{
custom_cluster_pathway_enrichment(
  cluster_res,
  pathway_member_list,
  source = "custom",
  feature_to_gene = MotrpacRatTraining6moData::FEATURE_TO_GENE_FILT,
  gene_identifier_type = "gene_symbol",
  universe = MotrpacRatTraining6moData::GENE_UNIVERSES$gene_symbol,
  add_ensembl_intersection = TRUE,
  min_input_set_size = 1,
  min_pw_set_size = 10,
  max_pw_set_size = 200,
  adjust_p = TRUE,
  num_cores = NULL,
  logfile = "/dev/null"
)
}
\arguments{
\item{cluster_res}{Either a data frame or a list of lists.
If a data frame, it needs at least two columns: "feature" and "cluster".
The "feature" column should be in the format
'\link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV};\link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV};feature_ID'.
If a list of lists, each sublist must be named with the cluster name (character string),
and the values must be features in the format
'\link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV};\link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV};feature_ID'.}

\item{pathway_member_list}{named list of character vectors where names are pathway names
and values are pathway members. Pathway members must match values in the
\code{gene_identifer_type} column of \code{feature_to_gene}.}

\item{source}{optional character string to define the source of \code{pathway_member_list}}

\item{feature_to_gene}{data frame, map between \code{intersection_id_type} and gene symbols.
Columns must include "feature_ID", "gene_symbol", "ensembl_gene", and "kegg_id".
\link[MotrpacRatTraining6moData:FEATURE_TO_GENE_FILT]{MotrpacRatTraining6moData::FEATURE_TO_GENE_FILT} by default.}

\item{gene_identifier_type}{character, column in \code{feature_to_gene} that
matches the gene identifier type in \code{universe}.
"gene_symbol" by default.}

\item{universe}{list of lists of character vectors,
named first by assay (i.e., \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV})
and then by tissue (i.e., \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}).
Vectors provide the full set of gene symbols associated with features tested
during differential analysis. For example, \code{universe$TRNSCRPT$LUNG}
should be a character vector of expressed genes in the lung, where the type
of gene identifier matches \code{gene_identifier_type}.
\code{[MotrpacRatTraining6moData::GENE_UNIVERSES]$gene_symbol} by default.}

\item{add_ensembl_intersection}{bool, whether to add a \code{intersection_ensembl}
column, which converts gene IDs in the intersection to Ensembl IDs}

\item{min_input_set_size}{integer, input must have this minimum number of
unique mappable gene IDs to attempt enrichment}

\item{min_pw_set_size}{integer, pathway must have at least this many members
to attempt enrichment}

\item{max_pw_set_size}{integer, pathway must have no more than this many members
to attempt enrichment}

\item{adjust_p}{boolean, whether to adjust nominal p-values for multiple testing (IHW by tissue)}

\item{num_cores}{optional integer, number of cores to register if parallel computing is desired}

\item{logfile}{optional character, path to log of failed iterations}
}
\value{
data frame with enrichment results, or NULL if no enrichment results were returned:
\describe{
\item{\code{term_size}}{integer, number of genes that are annotated to the term}
\item{\code{query_size}}{integer, number of genes that were included in the query}
\item{\code{intersection_size}}{integer, the number of genes in the input query that are annotated to the corresponding term}
\item{\code{term_id}}{character, unique term/pathway identifier}
\item{\code{source}}{character, the abbreviation of the data source for the term/pathway}
\item{\code{term_name}}{character, term/pathway name}
\item{\code{effective_domain_size}}{integer, the total number of genes in the universe used for the hypergeometric test}
\item{\code{intersection}}{character, input gene IDs that intersect with the term/pathway}
\item{\code{computed_p_value}}{double, nominal hypergeometric p-value}
\item{\code{cluster}}{character, cluster label}
\item{\code{tissue}}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}
\item{\code{ome}}{character, assay abbreviation, one of \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV}}
\item{\code{adj_p_value}}{double, adjusted nominal p-value
\code{computed_p_value} using IHW with tissue as a covariate}
}
}
\description{
Wrapper for multi-tissue, multi-omic pathway enrichment of clustering or
graphical results with a user-supplied list of pathways.
Pathway enrichment is performed using a hypergeometric test separately for
each unique combination of tissue, assay/ome, and cluster.
}
\examples{
# Use graphical clusters as an example
cluster_res = extract_main_clusters()
# Pick a single graphical cluster
# Gastrocnemius features up-regulated in both males and females at 8 weeks of training
cluster_res = cluster_res[cluster_res$cluster == "SKM-GN:8w_F1_M1",]

# Make a toy pathway member list with human gene symbols
pathways = list("TCA cycle" = c('ACO2','CS','FH','MDH1','OGDH','PDHA1','PDHA2','SDHC','SUCLG1'))

# Convert human gene symbols to rat gene symbols 
data("RAT_TO_HUMAN_GENE", package = "MotrpacRatTraining6moData")
for (pw in names(pathways)){
  newmembers = c()
  for (m in pathways[[pw]]){
    # get rat symbol
    rat = RAT_TO_HUMAN_GENE$RAT_SYMBOL[RAT_TO_HUMAN_GENE$HUMAN_ORTHOLOG_SYMBOL == m][1]
    if(!is.na(rat)){
      newmembers = c(newmembers, rat)
    }
  }
  pathways[[pw]] = newmembers
}

# Perform pathway enrichment 
custom_cluster_pathway_enrichment(cluster_res, 
                                  pathway_member_list = pathways,
                                  min_pw_set_size = 1)

}
\seealso{
\code{\link[=pathway_hypergeom_test]{pathway_hypergeom_test()}}, \code{\link[=cluster_pathway_enrichment]{cluster_pathway_enrichment()}}
}
