% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_to_numeric}
\alias{df_to_numeric}
\title{Make data frame numeric only}
\usage{
df_to_numeric(df, rownames = "feature_ID")
}
\arguments{
\item{df}{data frame with at least one numeric column}

\item{rownames}{either NULL, a character specifying a column name in \code{df},
or a vector. Specifies target row names. Default: "feature_ID"}
}
\value{
numeric data frame
}
\description{
Set row names and remove all non-numeric columns. This is useful for
reformatting data objects in \code{MotrpacRatTraining6moData},
e.g., \link[MotrpacRatTraining6moData:PROT_NORM_DATA]{MotrpacRatTraining6moData::PROT_HEART_NORM_DATA}
}
\examples{
df = MotrpacRatTraining6moData::PROT_HEART_NORM_DATA
df[1:5,1:8]
df_to_numeric(df)[1:5,1:4]

df = load_sample_data("SKM-GN", "TRNSCRPT")
df[1:5,1:8]
df_to_numeric(df)[1:5,1:4]

df = MotrpacRatTraining6moData::METAB_NORM_DATA_FLAT 
df[1:5,1:8]
rn = paste(df$assay, df$tissue, df$feature_ID, df$dataset, sep=";")
df_to_numeric(df, rownames = rn)[1:5,1:3]

}
