% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_datasets.R
\name{dl_format_pheno}
\alias{dl_format_pheno}
\title{Retrieve phenotypic data}
\usage{
dl_format_pheno(scratch, gsutil_path, parallel = TRUE)
}
\arguments{
\item{scratch}{directory in which to download files}

\item{gsutil_path}{path to \code{gsutil} binary. If you aren't sure where this is, run \code{which gsutil} in the Terminal and copy the result.}

\item{parallel}{whether or not to re-download the file from GCP if it already exists locally}
}
\value{
data.frame with formatted phenotypic data; row names are sample identifiers (vial labels)
}
\description{
Download and format phenotypic data from the endurance exercise training study in 6-month-old rats.
Codes are replaced with human-readable strings. 
Note that this currently does not handle variables that are recorded multiple times.
}
\examples{
dl_format_pheno("~/Documents/motrpac_data", "/usr/local/bin/google-cloud-sdk/bin/gsutil")
dl_format_pheno("~/Documents/motrpac_data", "/usr/local/bin/google-cloud-sdk/bin/gsutil", parallel = FALSE)

}
