% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp.R
\name{dl_read_gcp}
\alias{dl_read_gcp}
\title{Read a single file from Google Cloud into a data.table}
\usage{
dl_read_gcp(
  path,
  sep = "\\t",
  tmpdir = "/oak/stanford/groups/smontgom/nicolerg/tmp",
  GSUTIL_PATH = "~/google-cloud-sdk/bin/gsutil",
  check_first = F
)
}
\arguments{
\item{path}{GCP path, i.e. starts with "gs://"}

\item{sep}{column separator to use with "fread"}

\item{tmpdir}{scratch path to download files from GCP}

\item{GSUTIL_PATH}{path to "gsutil" on your computer}

\item{check_first}{check if file exists before downloading it. read in existing file if it exists. should be set to TRUE if you are running this function in parallel}
}
\value{
A data.table
}
\description{
Read a single file from Google Cloud into a data.table
}
