% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get}
\alias{.get}
\title{Get object from MotrpacRatTraining6moData}
\usage{
.get(object_name_as_string)
}
\arguments{
\item{object_name_as_string}{character, name of data object in \code{MotrpacRatTraining6moData}
R package}
}
\value{
specified object
}
\description{
For internal use only. Users can use \code{get()} or \code{data()} directly
after attaching the \code{MotrpacRatTraining6moData} package.
}
\keyword{internal}
