% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment_networks.R
\name{edge_intersection}
\alias{edge_intersection}
\title{Return intersection of gene symbols}
\usage{
edge_intersection(symbol1, symbol2)
}
\arguments{
\item{ensembl1}{character, comma-separated gene symbols of intersection from start node}

\item{ensembl2}{character, comma-separated gene symbols of intersection from end node}
}
\value{
string of comma-separated gene symbols in intersection
}
\description{
Function used internally in \code{\link[=enrichment_network_vis]{enrichment_network_vis()}}.
}
\seealso{
\code{\link[=enrichment_network_vis]{enrichment_network_vis()}}
}
