% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment_networks.R
\name{enrichment_network_vis}
\alias{enrichment_network_vis}
\title{Pathway enrichment network}
\usage{
enrichment_network_vis(
  pw_enrich_res = NULL,
  tissues = NULL,
  assays = MotrpacRatTraining6moData::ASSAY_ABBREV,
  cluster = NULL,
  feature_to_gene = MotrpacRatTraining6moData::FEATURE_TO_GENE,
  intersection_id_type = "ensembl_gene",
  similarity_cutoff = 0.375,
  adj_pval_cutoff = 0.1,
  title = NULL,
  label_nodes = TRUE,
  add_group_label_nodes = FALSE,
  parent_pathways = MotrpacRatTraining6moData::PATHWAY_PARENTS,
  multitissue_pathways_only = FALSE,
  include_metab_singletons = TRUE,
  similarity_scores_file = NULL,
  out_html = "/dev/null",
  overwrite_html = TRUE,
  return_html = FALSE,
  return_graph_for_cytoscape = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{pw_enrich_res}{optional data frame of enrichment results.
May be a subset of \link[MotrpacRatTraining6moData:GRAPH_PW_ENRICH]{MotrpacRatTraining6moData::GRAPH_PW_ENRICH}
or a data frame returned from \code{\link[=cluster_pathway_enrichment]{cluster_pathway_enrichment()}}.
Columns must include "adj_p_value", "ome", "tissue", "intersection", "computed_p_value",
"term_name", "term_id". If not specified, \code{tissues} and \code{cluster} must be specified.
\code{NULL} by default.}

\item{tissues}{character vector of tissue abbreviations, at least one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}.
Used to subset \link[MotrpacRatTraining6moData:GRAPH_PW_ENRICH]{MotrpacRatTraining6moData::GRAPH_PW_ENRICH}.
Must be specified if \code{pw_enrich_res} is not provided. \code{NULL} by default.}

\item{assays}{optional character vector of assay abbreviations, at least one of \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV}.
Used to subset \link[MotrpacRatTraining6moData:GRAPH_PW_ENRICH]{MotrpacRatTraining6moData::GRAPH_PW_ENRICH}.
All assays by default.}

\item{cluster}{character, graphical cluster of interest.
Used to subset \link[MotrpacRatTraining6moData:GRAPH_PW_ENRICH]{MotrpacRatTraining6moData::GRAPH_PW_ENRICH}.
Must be specified if \code{pw_enrich_res} is not provided.
e.g., "8w_F1_M1" for features that are
up-regulated (1) in females (F) and up-regulated in males (M1) at 8 weeks of training (8w), or
"1w_F0_M1->2w_F0_M1->4w_F0_M1->8w_F1_M1" for features that are up-regulated in males at all time points
(1w_F?_M1, 2w_F?_M1, 4w_F?_M1, 8w_F?_M1) and up-regulated in females at 8 weeks of training only.
See \link[MotrpacRatTraining6moData:GRAPH_PW_ENRICH]{MotrpacRatTraining6moData::GRAPH_PW_ENRICH}\code{$graphical_cluster} for all possible options.
See the \href{https://motrpac.github.io/MotrpacRatTraining6mo/articles/MotrpacRatTraining6mo.html#bayesian-graphical-clustering}{vignette}
for a detailed description of how to interpret and construct these strings. \code{NULL} by default.}

\item{feature_to_gene}{data frame, map between \code{intersection_id_type} and gene symbols.
Columns must include "gene_symbol" and "ensembl_gene" if \code{intersection_id_type == "ensembl_gene"}.
\link[MotrpacRatTraining6moData:FEATURE_TO_GENE]{MotrpacRatTraining6moData::FEATURE_TO_GENE} by default.}

\item{intersection_id_type}{character, type of gene identifier used to define
\code{pw_enrich_res$intersection}, either "ensembl_gene" or "gene_symbol".
"ensembl_gene" by default.}

\item{similarity_cutoff}{numeric, edges are drawn between pairs of pathways if they have
a similarity metric above this value. 0.375 by default.}

\item{adj_pval_cutoff}{numeric, pathway enrichments are only considered for
the network if the corresponding adjusted p-value (\code{pw_enrich_res$adj_p_value})
is less than this value. 0.1 (10\% FDR) by default.}

\item{title}{character, plot title. \code{NULL} by default.}

\item{label_nodes}{boolean, whether to label groups at all. \code{TRUE} by default.
If \code{label_nodes = FALSE}, \code{add_group_label_nodes} is ignored.}

\item{add_group_label_nodes}{boolean, whether to label groups with nodes.
If \code{FALSE}, use a standard legend instead.}

\item{parent_pathways}{named list, map of KEGG and REAC pathway term ID to parent pathway.
Used to create labels for clusters of pathway enrichments.
List names must correspond to values in \code{pw_enrich_res$term_id}.
\link[MotrpacRatTraining6moData:PATHWAY_PARENTS]{MotrpacRatTraining6moData::PATHWAY_PARENTS} by default.
If NULL, then only the pathway names are used to create cluster labels, which
is quite meaningless is all pathway names are unique. In that case,
it is recommended to set \code{label_nodes} to \code{FALSE}.}

\item{multitissue_pathways_only}{boolean. If \code{TRUE}, only include pathways
in the network if they are significantly enriched in more than one tissue
in \code{pw_enrich_res}.}

\item{include_metab_singletons}{boolean. If \code{TRUE}, include pathways enriched only
by metabolites (ome METAB) as singleton nodes}

\item{similarity_scores_file}{character or \code{NULL}, path in which to save
pairwise pathway similarity scores as an RDS file. If this file exists, the pairwise
similarity scores are loaded from the file instead of being recalculated.
\code{NULL} by default.}

\item{out_html}{character, output file for HTML. "/dev/null" by default.}

\item{overwrite_html}{boolean, whether to overwrite \code{out_html} if it already exists.
\code{TRUE} by default. If \code{out_html} exists and \code{overwrite_html=FALSE},
then the path \code{out_html} is returned.}

\item{return_html}{boolean, whether to return the path to \code{out_html}. If
\code{FALSE}, return the \code{visNetwork} object instead. \code{FALSE} by default.}

\item{return_graph_for_cytoscape}{boolean. If \code{TRUE}, return an \code{igraph}
graph object that can be exported to Cytoscape.}

\item{verbose}{boolean, whether to print verbose output}
}
\value{
If \code{return_graph_for_cytoscape=FALSE} and \code{return_html=FALSE}, return a
\code{visNetwork} graph. If \code{return_graph_for_cytoscape=TRUE}, return an
igraph object. If \code{return_html=TRUE}, return the path to the HTML file in
which the interactive \code{visNetwork} was saved.
}
\description{
Plot an interactive network of pathway enrichments.
}
\examples{
# Example 1: Plot an interactive network of pathway enrichments from the HEART:8w_F1_M1 node,
# i.e., features that are up-regulated in both males and females after 8 weeks of training
enrich_res = MotrpacRatTraining6moData::GRAPH_PW_ENRICH
enrich_res = enrich_res[enrich_res$cluster == "HEART:8w_F1_M1",]
enrichment_network_vis(enrich_res, add_group_label_nodes = TRUE) 
# Equivalently:
enrichment_network_vis(tissues="HEART", 
                       cluster="8w_F1_M1",
                       add_group_label_nodes = TRUE) 

\dontrun{
# Example 2: Export the above network to Cytoscape.
# Cytoscape must be running locally for this to work. 
library(RCy3)
g = enrichment_network_vis(enrich_res, return_graph_for_cytoscape = T) 
RCy3::cytoscapePing()
RCy3::createNetworkFromIgraph(g, new.title='HEART:8w_F1_M1')
}

# Example 3: Plot an interactive network of pathway enrichments corresponding to 
# features that are up-regulated in both sexes at 8 weeks in any of the 3 muscle tissues. 
# Only include pathways that are enriched in at least 2 muscle tissues.
# Plot the color legend on the left instead of using colored labels pointing to nodes 
enrich_res = MotrpacRatTraining6moData::GRAPH_PW_ENRICH
enrich_res = enrich_res[enrich_res$tissue \%in\% c("SKM-GN","SKM-VL","HEART") &
                         grepl(":8w_F1_M1",enrich_res$cluster),]
enrichment_network_vis(enrich_res,
                       similarity_cutoff = 0.3,
                       title = "Muscle 8w_F1_M1",
                       add_group_label_nodes = FALSE,
                       multitissue_pathways_only = TRUE) 
# Equivalently:
enrichment_network_vis(tissues = c("SKM-GN","SKM-VL","HEART"),
                       cluster = "8w_F1_M1",
                       similarity_cutoff = 0.3,
                       title = "Muscle 8w_F1_M1",
                       add_group_label_nodes = FALSE,
                       multitissue_pathways_only = TRUE) 

}
\seealso{
Ancillary functions include \code{\link[=calc_similarity_metric]{calc_similarity_metric()}}, \code{\link[=replace_ensembl_with_symbol]{replace_ensembl_with_symbol()}},
\code{\link[=format_gene_symbols]{format_gene_symbols()}}, \code{\link[=edge_intersection]{edge_intersection()}}, \code{\link[=collapse_p]{collapse_p()}}, \code{\link[=add_line_breaks]{add_line_breaks()}},
\code{\link[=format_gene_lists]{format_gene_lists()}}, and \code{\link[=cleanup]{cleanup()}}.
The input \code{pw_enrich_res} can also be generated with \code{\link[=cluster_pathway_enrichment]{cluster_pathway_enrichment()}}.
}
