% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_graphical_clustering.R
\name{extract_tissue_sets}
\alias{extract_tissue_sets}
\title{An auxiliary function for getting the top analyte sets for a set of tissues.}
\usage{
extract_tissue_sets(
  tissues,
  node_sets,
  edge_sets,
  k = 3,
  min_size = 20,
  add_week8 = T,
  omes = NULL
)
}
\arguments{
\item{tissues}{A character vector. The names of the tissues (one or more) to be considered.}

\item{node_sets}{A named list with the node (state) sets of analytes/features.}

\item{edge_sets}{A named list with the edge (state) sets of analytes/features.}

\item{k}{An integer. How many node/edge/trajectory sets to extract. See details.}

\item{add_week8}{A logical. TRUE (default): add all week 8 nodes to the node set.}

\item{omes}{A character vector (optional). The names of the omes (one or more) to be considered.}
}
\value{
A named list. Names correspond to set names (can correspond to a node, edge, or a trajectory).
}
\description{
An auxiliary function for getting the top analyte sets for a set of tissues.
}
\details{
This function is useful for extracting the largest sets for a specific set of tissues and omes.
Thus, it is a useful step before running enrichment analysis on sets of analytes identified by the graphical clustering analysis.

By specifying k the user can control how many sets to include. The default is three, which means that the top largest node sets, edge sets, and full trajectories will be extracted (three each).

Naming format:
Analyte names are in the ome;tissue;feature_id format.
Node set names: {week}_F{differential regulation status}_M{differential regulation status}.
Example 1: 1w_F1_M0 means up-regulation in females week 1 and null (zero effect) in males in week 1.
Example 2: 1w_F1_M-1 means up-regulation in females week 1 and down-regulation in males in week 1.
The edge set object is a named list of string vectors. The name of an edge is node_id---node_id.
A full trajectory is annotated by node_id->node_id->node_id->node_id. For example: 1w_F-1_M-1->2w_F-1_M-1->4w_F-1_M-1->8w_F-1_M-1.

See bayesian_graphical_clustering for more details about the graphical analysis.
}
\examples{
data(GRAPH_COMPONENTS)
filtered_solution = extract_tissue_sets(
   tissues = c("HEART","SKM-VL","SKM-GN"),
   node_sets=GRAPH_COMPONENTS$node_sets,
   edge_sets=GRAPH_COMPONENTS$edge_sets
)
}
