% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_graphical_clustering.R
\name{extract_top_trajectories}
\alias{extract_top_trajectories}
\title{Extract top trajectories}
\usage{
extract_top_trajectories(
  tissues,
  omes = MotrpacRatTraining6moData::ASSAY_ABBREV,
  k = 5,
  min_size = 5,
  add_week8 = FALSE,
  node_sets = MotrpacRatTraining6moData::GRAPH_COMPONENTS$node_sets,
  edge_sets = MotrpacRatTraining6moData::GRAPH_COMPONENTS$edge_sets
)
}
\arguments{
\item{tissues}{character, tissue abbreviation, at least one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}

\item{omes}{character, assay abbreviation, one of \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV}. All assays by default.}

\item{k}{integer, return the top \code{k} largest trajectories/paths with at least \code{min_size}
features. Default: 5}

\item{min_size}{integer, minimal cluster size to be considered. Clusters with fewer than
this number of features are excluded. Default: 5}

\item{add_week8}{logical, whether to include non-null 8-week nodes with at least
\code{min_size} features in addition to the top \code{k} largest trajectories/paths.
Default: FALSE}

\item{node_sets}{named list with the node (state) sets of analytes/features.
\code{\link[MotrpacRatTraining6moData]{GRAPH_COMPONENTS}$node_sets} by default.}

\item{edge_sets}{named list with the edge (state) sets of analytes/features.
\code{\link[MotrpacRatTraining6moData]{GRAPH_COMPONENTS}$edge_sets} by default.}
}
\value{
named list where names are the names of the graphical clusters (paths/trajectories and,
optionally, non-null 8-week nodes) and values are vectors of the training-regulated
features that belong to that cluster, in the format
\code{\link[MotrpacRatTraining6moData]{ASSAY_ABBREV};\link[MotrpacRatTraining6moData]{TISSUE_ABBREV};[feature_ID]}.
}
\description{
Get the features corresponding to each of the top k largest trajectories of training-regulated
features for the given tissue(s) and ome(s). Optionally include features in the non-null
8-week nodes as well (e.g., 8w_F1_M1, 8w_F0_M-1).
}
\examples{
# Top 3 trajectories of training-regulated proteins in the liver
res = extract_top_trajectories("LIVER", omes="PROT")
names(res)
lapply(res, length)

# Top 5 trajectories and 8-week nodes for the union of training-regulated 
# features in the heart and gastrocnemius (all omes)
res = extract_top_trajectories(c("HEART","SKM-GN"), add_week8=TRUE)
names(res)
lapply(res, length)
}
