% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetch_object}
\alias{fetch_object}
\title{Get object from MotrpacRatTraining6moData}
\usage{
fetch_object(name_as_string)
}
\arguments{
\item{name_as_string}{character, name of object in \code{MotrpacRatTraining6moData}}
}
\value{
specified object
}
\description{
Using \code{get()} to retrieve data from MotrpacRatTraining6moData
within this package doesn't work, at least in tests.
This function allows us to be explicit about the source package.
For internal use only. Users can use \code{get()} or \code{data()} directly
after loading the \code{MotrpacRatTraining6moData} package.
}
\seealso{
\code{\link[=load_sample_data]{load_sample_data()}}
}
\keyword{internal}
