% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{find_flanks}
\alias{find_flanks}
\title{Find human flanks}
\usage{
find_flanks(x, fasta, type = "uniprot")
}
\arguments{
\item{x}{character, human accession for PTM, e.g., "Q96QG7_S548s"}

\item{fasta}{\link[Biostrings:XStringSet-class]{Biostrings::XStringSet} object returned from reading in
a human protein FASTA file with \code{\link[Biostrings:XStringSet-io]{Biostrings::readAAStringSet()}}. Names of the
\link[Biostrings:XStringSet-class]{Biostrings::XStringSet} object should be set to the human protein accession,
e.g., "Q96QG7". Can be the result of \code{\link[=load_uniprot_human_fasta]{load_uniprot_human_fasta()}}, which
returns the version of the Uniprot human protein FASTA used in the manuscript.}

\item{type}{character, type of accession}
}
\value{
character, pipe-separated string of human flanking sequences for the PTM
}
\description{
Find flanking sequence in human ortholog of PTM
}
\examples{
fasta = load_uniprot_human_fasta(scratchdir="/tmp")
x = "Q96QG7_S548s"
find_flanks(x, fasta)
}
\seealso{
\code{\link[=load_uniprot_human_fasta]{load_uniprot_human_fasta()}}
}
