% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_meta_regression.R
\name{forest_plot}
\alias{forest_plot}
\title{Print forest plot}
\usage{
forest_plot(results, metabolite_refmet)
}
\arguments{
\item{results}{named list returned by \code{\link[=metab_meta_analysis]{metab_meta_analysis()}} or \code{\link[=metab_meta_regression]{metab_meta_regression()}}}

\item{metabolite_refmet}{character, RefMet name of metabolite}
}
\description{
Make a forest plot to present multiple measurements and consensus effect
size from meta-analysis or meta-regression. Note that using \code{\link[=metab_meta_regression]{metab_meta_regression()}}
results as input generates a single forest plot while using \code{\link[=metab_meta_analysis]{metab_meta_analysis()}}
generates one forest plot per sex and training time point.
}
\examples{
# Get meta-analysis results in gastrocnemius
res = metab_meta_analysis("SKM-GN")
# Pick a feature
res$merged_res[res$merged_res$min_nominal_p > 0.01 & 
  res$merged_res$p_value < 0.001 & 
  res$merged_res$I2 < 30,]
forest_plot(res, metabolite_refmet="Hydroxyproline")

# Look at meta-regression results for a feature in the plasma
res = metab_meta_regression("PLASMA")
forest_plot(res, metabolite_refmet="CAR(18:2)")

res = metab_meta_regression("BAT")
forest_plot(res, metabolite_refmet="Aspartic acid")
}
\seealso{
\code{\link[=metab_meta_analysis]{metab_meta_analysis()}}, \code{\link[=metab_meta_regression]{metab_meta_regression()}}
}
