% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment_networks.R
\name{format_gene_lists}
\alias{format_gene_lists}
\title{Format gene lists}
\usage{
format_gene_lists(x)
}
\arguments{
\item{x}{value in \code{points$enriched} column}
}
\value{
character, new value with additional line breaks
}
\description{
Apply \code{\link[=add_line_breaks]{add_line_breaks()}} to the \code{points$enriched} column.
Function is used internally in \code{\link[=enrichment_network_vis]{enrichment_network_vis()}}.
}
\seealso{
\code{\link[=enrichment_network_vis]{enrichment_network_vis()}}, \code{\link[=add_line_breaks]{add_line_breaks()}}
}
