% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment_networks.R
\name{format_gene_symbols}
\alias{format_gene_symbols}
\title{Format gene symbols}
\usage{
format_gene_symbols(x, return_N = TRUE, collapse = TRUE)
}
\arguments{
\item{x}{character, comma-separated gene symbols}

\item{return_N}{boolean, whether to prepend the concatenated
gene symbols with the number of unique gene symbols.
\code{TRUE} by default.}

\item{collapse}{boolean, whether to collapse the gene symbols into a
comma-separated string. If not, return a vector of unique gene symbols.
\code{TRUE} by default.}
}
\value{
either a string of comma-separated gene symbols,
a vector of unique gene symbols,
or a named list with two values ("N" and "genes")
depending on the values of \code{return_N} and \code{collapse}.
}
\description{
Format a string of comma-separated gene symbols for use in \code{\link[=enrichment_network_vis]{enrichment_network_vis()}}.
}
\seealso{
\code{\link[=enrichment_network_vis]{enrichment_network_vis()}}
}
\keyword{internal}
