% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment.R
\name{gene_pathway_enrichment}
\alias{gene_pathway_enrichment}
\title{Gene pathway enrichment}
\usage{
gene_pathway_enrichment(
  input,
  background,
  organism = "rnorvegicus",
  databases = c("REAC", "WP", "KEGG"),
  min_pw_set_size = 10,
  max_pw_set_size = 200,
  return_gem = FALSE
)
}
\arguments{
\item{input}{vector of gene identifiers}

\item{background}{vector of gene identifiers in universe/background. Must include the
input. Should include all genes that were candidates for \code{input}, e.g.,
expressed genes. If performing pathway enrichment on genes identified through \code{MotrpacRatTraining6moData},
see \link[MotrpacRatTraining6moData:GENE_UNIVERSES]{MotrpacRatTraining6moData::GENE_UNIVERSES}. See \code{custom_bg} argument for \code{\link[gprofiler2:gost]{gprofiler2::gost()}}.}

\item{organism}{character, species name. Default: "rnorvegicus" See \code{organism}
argument for \code{\link[gprofiler2:gost]{gprofiler2::gost()}}.}

\item{databases}{character, vector of databases in which to search. See \code{sources}
argument for \code{\link[gprofiler2:gost]{gprofiler2::gost()}}.}

\item{min_pw_set_size}{integer, exclude pathways smaller than this. Default: 10}

\item{max_pw_set_size}{integer, exclude pathways larger than this. Default: 200}

\item{return_gem}{bool, whether to return a data frame compatible with the
Generic Enrichment Map (GEM) file format, which can be used as an input for the
\href{https://apps.cytoscape.org/apps/enrichmentmap}{Cytoscape EnrichmentMap application}.
Default: FALSE}
}
\value{
a data frame of pathway enrichment results. If \code{return_gem=TRUE},
column names are changed add added to be compatible with the Generic Enrichment Map (GEM) file format.
Otherwise, the columns are defined as follows:
\describe{
\item{\code{query}}{character, the name of the input query which by default is the order of query with the prefix "query_"}
\item{\code{significant}}{bool, whether the \code{\link[gprofiler2:gost]{gprofiler2::gost()}} enrichment p-value is less than 0.05}
\item{\code{term_size}}{integer, number of genes that are annotated to the term}
\item{\code{query_size}}{integer, number of genes that were included in the query (from \code{\link[gprofiler2:gost]{gprofiler2::gost()}})}
\item{\code{intersection_size}}{integer, the number of genes in the input query that are annotated to the corresponding term}
\item{\code{precision}}{double, the proportion of genes in the input list that are annotated to the function, defined as \code{intersection_size/query_size}}
\item{\code{recall}}{double, the proportion of functionally annotated genes that the query recovers, defined as \code{intersection_size/term_size}}
\item{\code{term_id}}{character, unique term/pathway identifier}
\item{\code{source}}{character, the abbreviation of the data source for the term/pathway}
\item{\code{term_name}}{character, term/pathway name}
\item{\code{effective_domain_size}}{integer, the total number of genes in the universe used for the hypergeometric test}
\item{\code{source_order}}{integer, numeric order for the term within its data source}
\item{\code{parents}}{list of term IDs that are hierarchically directly above the term.
For non-hierarchical data sources this points to an artificial root node}
\item{\code{evidence_codes}}{character, comma-separated evidence codes}
\item{\code{intersection}}{character, input gene IDs that intersect with the term/pathway}
\item{\code{gost_adj_p_value}}{double, improperly adjusted hypergeometric p-value
from \code{\link[gprofiler2:gost]{gprofiler2::gost()}}. For reference only; should not be used to filter results
unless there was only a single list of genes in the input.}
\item{\code{computed_p_value}}{double, nominal hypergeometric p-value,
computed from the \code{\link[gprofiler2:gost]{gprofiler2::gost()}} output}
\item{\code{BH_adj_p_value}}{double, BH-adjusted p-values, calculated on \code{computed_p_value}}
}
}
\description{
Perform pathway enrichment for a list of genes.
Wrapper for \code{\link[gprofiler2:gost]{gprofiler2::gost()}}.
}
\examples{
# Perform pathway enrichment for differential transcripts in the liver
diff = MotrpacRatTraining6moData::TRAINING_REGULATED_FEATURES
input_feat = diff$feature_ID[diff$tissue == "LIVER" & diff$assay == "TRNSCRPT"]
map = MotrpacRatTraining6moData::FEATURE_TO_GENE_FILT
input = unique(map$gene_symbol[map$feature_ID \%in\% input_feat])
background = MotrpacRatTraining6moData::GENE_UNIVERSES$gene_symbol$TRNSCRPT$LIVER
res = gene_pathway_enrichment(input, background)
head(res)
}
