% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_graphical_clustering.R
\name{get_all_trajectories}
\alias{get_all_trajectories}
\title{Remove all non-empty trajectories}
\usage{
get_all_trajectories(
  edge_sets = MotrpacRatTraining6moData::GRAPH_COMPONENTS$edge_sets,
  node_sets = MotrpacRatTraining6moData::GRAPH_COMPONENTS$node_sets,
  tissues = MotrpacRatTraining6moData::TISSUE_ABBREV,
  omes = MotrpacRatTraining6moData::ASSAY_ABBREV
)
}
\arguments{
\item{edge_sets}{A named list with the edge (state) sets of analytes/features.
\code{\link[MotrpacRatTraining6moData]{GRAPH_COMPONENTS}$edge_sets} by default.}

\item{node_sets}{A named list with the node (state) sets of analytes/features.
\code{\link[MotrpacRatTraining6moData]{GRAPH_COMPONENTS}$node_sets} by default.}

\item{tissues}{string vector, optional. tissue subset. all tissues by default}

\item{omes}{string vector, optional. ome subset. all omes by default}
}
\value{
named list with one element per trajectories. members are features in the path
}
\description{
Remove all non-empty trajectories
}
\examples{
# Get lists of features belonging to all trajectories in the liver
get_all_trajectories(tissues = "LIVER")

}
