% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_prep_data.R
\name{get_file_from_url}
\alias{get_file_from_url}
\title{Load file from GCS}
\usage{
get_file_from_url(url, scratchdir = ".", return_data_table = FALSE)
}
\arguments{
\item{url}{character, Google Cloud Storage URL, e.g., "https://storage.googleapis.com/motrpac-rat-training-6mo-extdata/metabolomics-training-da/pass1b-06_t31-plasma_metab_training-dea-fdr_20211006.txt"}

\item{scratchdir}{character, local directory in which to download data from
the web. Current working directory by default. Downloaded files are deleted before returning the result.}

\item{return_data_table}{bool, whether to return a data table. If FALSE, return a data frame. FALSE by default.}
}
\value{
data frame, or data table if \code{return_data_table = TRUE}
}
\description{
Function to download a text file from the web.
}
\examples{
# Get non-meta-analyzed metabolomics training differential analysis results for blood plasma 
res = get_file_from_url(paste0(c("https://storage.googleapis.com/motrpac-rat-training-6mo-extdata/",
  "metabolomics-training-da/pass1b-06_t31-plasma_metab_training-dea-fdr_20211006.txt"),collapse=""))
}
