% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_peak_annotations}
\alias{get_peak_annotations}
\title{Get genomic peak annotations}
\usage{
get_peak_annotations(
  counts,
  species = "Rattus norvegicus",
  release = 96,
  txdb = NULL
)
}
\arguments{
\item{counts}{data table or data frame with columns 'chrom','start','end', and 'feature_ID' OR
for ATAC only, data frame with row names as feature IDs in the format format 'chrom:start-end'}

\item{species}{character, scientific name of the organism for which to create the
Ensembl database with \code{\link[GenomicFeatures:makeTxDbFromEnsembl]{GenomicFeatures::makeTxDbFromEnsembl()}}}

\item{release}{integer, the Ensembl release to query.
If set to NA, the current release is used.}

\item{txdb}{optional \link[GenomicFeatures:TxDb-class]{GenomicFeatures::TxDb} object, in which case
the database is not regenerated}
}
\value{
data frame with formatted \code{\link[ChIPseeker:annotatePeak]{ChIPseeker::annotatePeak()}} output
and additional columns "custom_annotation" and "relationship_to_gene".
}
\description{
Get and fix peak annotations from \code{\link[ChIPseeker:annotatePeak]{ChIPseeker::annotatePeak()}}
}
\details{
"relationship_to_gene" is the shortest distance between the feature and the start or end of the closest gene.
It is 0 if the feature has any overlap with the gene.
"custom_annotation" fixes many issues with the \code{ChIPseeker} annotation (v1.22.1).
}
