% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_prep_data.R
\name{get_rdata_from_url}
\alias{get_rdata_from_url}
\title{Load RData from GCS}
\usage{
get_rdata_from_url(
  tissue = NULL,
  assay = NULL,
  suffix = NULL,
  scratchdir = ".",
  url = NULL,
  obj_name = NULL,
  nrows = Inf
)
}
\arguments{
\item{tissue}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}. Only used if \code{url} is NULL.}

\item{assay}{character, assay abbreviation, one of \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV}. Only used if \code{url} is NULL.}

\item{suffix}{character, object suffix. Only used if \code{url} is NULL.}

\item{scratchdir}{character, local directory in which to download data from
the web. Current working directory by default.}

\item{url}{character, RData URL. Optional if URL cannot be constructed from
\code{tissue}, \code{assay}, and \code{suffix}.}

\item{obj_name}{character, name of object saved within RData file. Only required
if more than one object is saved in the RData file.}

\item{nrows}{integer, number of rows to return. Defaults to Inf.}
}
\value{
object saved in the RData file, optionally specified by \code{obj_name}
}
\description{
Function to download RData from the web. Intended to download
epigenetic data from Google Cloud Storage but can be used for any public URL
that points to an RData file.
}
\examples{
\dontrun{
# return gastrocnemius chromatin accessibility differential analysis results 
data = get_rdata_from_url(tissue="SKM-GN", assay="ATAC", suffix="DA", scratchdir="/tmp")

# return brown adipose DNA methylation raw counts 
data = get_rdata_from_url(tissue="BAT", assay="METHYL", suffix="RAW_COUNTS", scratchdir="/tmp")

# return DNA methylation feature annotation
data = get_rdata_from_url(url="https://storage.googleapis.com/motrpac-rat-training-6mo-extdata/METHYL_FEATURE_ANNOT.rda", scratchdir="/tmp")

# return raw DNA methylation data for brown adipose
data = get_rdata_from_url(url="https://storage.googleapis.com/motrpac-rat-training-6mo-extdata/raw/RRBS/BAT_raw.RData", scratchdir="/tmp")
}
}
