% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{limma_res_extract_se}
\alias{limma_res_extract_se}
\title{Extract standard errors}
\usage{
limma_res_extract_se(limma_res, effect_col = "logFC", t_col = "t")
}
\arguments{
\item{limma_res}{data frame returned by \code{\link[limma:toptable]{limma::topTable()}}}

\item{effect_col}{character, name of column containing the effect size}

\item{t_col}{character, column name containing the t statistic}
}
\value{
numeric vector of standard errors for effect sizes
}
\description{
Extract standard errors from \code{limma} results
}
\keyword{internal}
