% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_available_data}
\alias{list_available_data}
\title{List available data}
\usage{
list_available_data(package = NULL)
}
\arguments{
\item{package}{optional string to specify a package}
}
\value{
character vector of names of data objects available to load
}
\description{
List available data, including lazily-loaded data. Useful to gather
split data frames.
}
\examples{
list_available_data()
list_available_data("MotrpacBicQC")
}
