% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_prep_data.R
\name{load_atac_feature_annotation}
\alias{load_atac_feature_annotation}
\title{Load ATAC feature annotation}
\source{
\url{https://storage.googleapis.com/motrpac-rat-training-6mo-extdata/ATAC_FEATURE_ANNOT.rda}
}
\usage{
load_atac_feature_annotation(scratchdir = ".")
}
\arguments{
\item{scratchdir}{character, local directory in which to download data from
Google Cloud Storage. Current working directory by default.}
}
\value{
data frame. See \link[MotrpacRatTraining6moData:ATAC_FEATURE_ANNOT]{MotrpacRatTraining6moData::ATAC_FEATURE_ANNOT} for details.
}
\description{
Load ATAC feature annotation
}
\examples{
\dontrun{
feature_annot = load_atac_feature_annotation("/tmp")
}

}
