% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_dropbox_rdata}
\alias{load_dropbox_rdata}
\title{Download an RData file from Dropbox}
\usage{
load_dropbox_rdata(target, tmpdir, redownload = FALSE)
}
\arguments{
\item{target}{character, public URL specifying an RData file ('.rda' suffix) on Dropbox}

\item{tmpdir}{character, local path in which to download \code{target}}

\item{redownload}{boolean, whether or not to download the file if it already exists in \code{tmpdir}}
}
\value{
object contained in \code{target}
}
\description{
Download an RData file from Dropbox
}
\examples{
target = "https://www.dropbox.com/s/rsenv5dgsx0to4i/test.rda?raw=1"
obj = load_dropbox_rdata(target, tmpdir)
obj = load_dropbox_rdata(target, tmpdir, redownload = T)
}
