% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_prep_data.R
\name{load_epigen_da}
\alias{load_epigen_da}
\title{Load epigenetic differential analysis results}
\usage{
load_epigen_da(tissue, assay, scratchdir = ".")
}
\arguments{
\item{tissue}{character, tissue abbreviation, one of "BAT", "HEART", "HIPPOC", "KIDNEY", "LIVER", "LUNG", "SKM-GN", "WAT-SC"}

\item{assay}{character, one of "METHYL" (RRBS) or "ATAC"}

\item{scratchdir}{character, local directory in which to download data from
Google Cloud Storage. Current working directory by default.}
}
\value{
data frame. See \link{ATAC_DA} and \link{METHYL_DA} for details.
}
\description{
Load ATAC or METHYL timewise differential analysis results for a tissue from
Google Cloud Storage. See \link{ATAC_DA} and \link{METHYL_DA}
for more details.
}
\examples{
\dontrun{
# Return ATAC-seq differential analysis results for gastrocnemius skeletal muscle
load_epigen_da("SKM-GN", "METHYL")
}
}
