% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_prep_data.R
\name{load_feature_annotation}
\alias{load_feature_annotation}
\title{Load feature annotation}
\usage{
load_feature_annotation(assay, scratchdir = ".")
}
\arguments{
\item{assay}{character, assay abbreviation, one of \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV}}

\item{scratchdir}{character, local directory in which to download data from
Google Cloud Storage. Current working directory by default. Not used
if \code{assay} is "TRNSCRPT", "IMMUNO", or "METAB".}
}
\value{
data frame with one row per feature with unique metadata
}
\description{
Load feature annotation
}
\examples{
head(load_feature_annotation("UBIQ", scratchdir="/tmp"))
head(load_feature_annotation("TRNSCRPT", scratchdir="/tmp"))
head(load_feature_annotation("METAB"))
head(load_feature_annotation("IMMUNO"))

}
\seealso{
\link[MotrpacRatTraining6moData:FEATURE_TO_GENE]{MotrpacRatTraining6moData::FEATURE_TO_GENE},
\link[MotrpacRatTraining6moData:FEATURE_TO_GENE_FILT]{MotrpacRatTraining6moData::FEATURE_TO_GENE_FILT},
\link[MotrpacRatTraining6moData:PROT_FEATURE_ANNOT]{MotrpacRatTraining6moData::PROT_FEATURE_ANNOT},
\link[MotrpacRatTraining6moData:PHOSPHO_FEATURE_ANNOT]{MotrpacRatTraining6moData::PHOSPHO_FEATURE_ANNOT},
\link[MotrpacRatTraining6moData:ACETYL_FEATURE_ANNOT]{MotrpacRatTraining6moData::ACETYL_FEATURE_ANNOT},
\link[MotrpacRatTraining6moData:UBIQ_FEATURE_ANNOT]{MotrpacRatTraining6moData::UBIQ_FEATURE_ANNOT},
\link[MotrpacRatTraining6moData:TRNSCRPT_FEATURE_ANNOT]{MotrpacRatTraining6moData::TRNSCRPT_FEATURE_ANNOT},
\link[MotrpacRatTraining6moData:ATAC_FEATURE_ANNOT]{MotrpacRatTraining6moData::ATAC_FEATURE_ANNOT},
\link[MotrpacRatTraining6moData:METHYL_FEATURE_ANNOT]{MotrpacRatTraining6moData::METHYL_FEATURE_ANNOT},
\link[MotrpacRatTraining6moData:METAB_FEATURE_ID_MAP]{MotrpacRatTraining6moData::METAB_FEATURE_ID_MAP}
}
