% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_prep_data.R
\name{load_metabolomics_da}
\alias{load_metabolomics_da}
\title{Load metabolomics differential analysis}
\usage{
load_metabolomics_da(tissue, type = "timewise", ...)
}
\arguments{
\item{tissue}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}

\item{type}{character, type of differential analysis summary statistics to return. One of 'timewise', 'training'.
Note that the adjusted p-value from \code{type="training"} is already include in \code{type="timewise"} as \code{selection_fdr}.}

\item{...}{additional arguments passed to \code{\link[=get_file_from_url]{get_file_from_url()}}}
}
\value{
data frame. For detailed column descriptions, see \link[MotrpacRatTraining6moData:METAB_DA]{MotrpacRatTraining6moData::METAB_DA}
or \href{https://docs.google.com/spreadsheets/d/10I4ofH__sMhlONrwdH8ms9S4qkXpQ_wXMjcCa1fBUyY/edit#gid=0}{this online spreadsheet}.
}
\description{
Load training or timewise metabolomics differential analysis results for a single tissue.
For timewise results, load from \link[MotrpacRatTraining6moData:METAB_DA]{MotrpacRatTraining6moData::METAB_DA}. For training results,
load from GCS.
These are the redundant, non-meta-analyzed differential analysis that were used as input for the
metabolomics meta-analysis and meta-regression but not otherwise directly used in the manuscript.
The differential analysis results provided by \link[MotrpacRatTraining6moData:METAB_DA_METAREG]{MotrpacRatTraining6moData::METAB_DA_METAREG} are the
non-redundant, meta-regression analysis results used in the manuscript.
}
\examples{
# Get timewise differential analysis results for the liver
res = load_metabolomics_da("LIVER")

# Get training differential analysis results for the liver
res = load_metabolomics_da("LIVER", type="training")

}
\seealso{
\code{\link[=metabolomics_meta_analysis]{metabolomics_meta_analysis()}}, \code{\link[=metabolomics_meta_regression]{metabolomics_meta_regression()}}, \link[MotrpacRatTraining6moData:METAB_DA]{MotrpacRatTraining6moData::METAB_DA}
}
