% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_prep_data.R
\name{load_methyl_feature_annotation}
\alias{load_methyl_feature_annotation}
\title{Load METHYL feature annotation}
\source{
\url{https://storage.googleapis.com/motrpac-rat-training-6mo-extdata/epigen-rda/METHYL_FEATURE_ANNOT.rda}
}
\usage{
load_methyl_feature_annotation(scratchdir = ".")
}
\arguments{
\item{scratchdir}{character, local directory in which to download data from
Google Cloud Storage. Current working directory by default.}
}
\value{
data frame. See \link[MotrpacRatTraining6moData:METHYL_FEATURE_ANNOT]{MotrpacRatTraining6moData::METHYL_FEATURE_ANNOT} for details.
}
\description{
Load METHYL feature annotation from Google Cloud Storage.
See \link{METHYL_FEATURE_ANNOT} for more details.
}
\examples{
\dontrun{
feature_annot = load_methyl_feature_annotation("/tmp")
}

}
