% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_prep_data.R
\name{load_sample_data}
\alias{load_sample_data}
\title{Load sample-level data}
\usage{
load_sample_data(
  tissue,
  assay,
  normalized = TRUE,
  training_regulated_only = FALSE,
  exclude_outliers = FALSE,
  scratchdir = ".",
  nrows = Inf,
  warnings = TRUE
)
}
\arguments{
\item{tissue}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}

\item{assay}{character, assay abbreviation, one of \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV}}

\item{normalized}{bool, whether to return normalized data. If \code{FALSE}, return raw counts.}

\item{training_regulated_only}{bool, whether to filter features down to those training-regulated at 5\% FDR}

\item{exclude_outliers}{bool, whether to remove sample outliers specified by \link[MotrpacRatTraining6moData:OUTLIERS]{MotrpacRatTraining6moData::OUTLIERS}}

\item{scratchdir}{character, local directory in which to download data from
Google Cloud Storage. Current working directory by default. Only applies if \code{assay} is ATAC or METHYL.}

\item{nrows}{integer, number of rows to return. Defaults to Inf. Useful to return a subset of a large data frame for tests.}

\item{warnings}{bool, whether to print warnings to the console. \code{TRUE} by default.}
}
\value{
a data.frame where features are in rows and numeric columns correspond to sample identifiers (vial labels)
}
\description{
Load raw counts or normalized sample-level data. Optionally filter by training-regulated features.
For epigenetic data (ATAC and METHYL), if \code{training_regulated_only = FALSE}, sample-level data
is downloaded from Google Cloud Storage.
}
\examples{
# Load RNA-seq raw counts for liver
data = load_sample_data("LIVER", "TRNSCRPT", normalized = FALSE)

# Load normalized metabolomics data for gastrocnemius
data = load_sample_data("SKM-GN", "METAB")

# Load normalized protein abundance data for heart
data = load_sample_data("HEART", "PROT")

\dontrun{
# Load ATAC-seq raw counts for hippocampus, excluding outliers 
data = load_sample_data("HIPPOC", 
                        "ATAC", 
                        exclude_outliers = TRUE, 
                        normalized = FALSE, 
                        scratchdir = "/tmp")
}
}
\seealso{
\code{\link[=combine_normalized_data]{combine_normalized_data()}}
}
