% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_prep_data.R
\name{load_training_da}
\alias{load_training_da}
\title{Load training differential analysis results}
\usage{
load_training_da(assay, tissue, metareg = TRUE, ...)
}
\arguments{
\item{assay}{character, assay abbreviation, one of \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV}}

\item{tissue}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}

\item{metareg}{bool, whether to return the meta-regression results for METAB.
Return the upstream differential analysis results if \code{FALSE}.
\code{TRUE} by default.}

\item{...}{additional arguments passed to \code{\link[=get_file_from_url]{get_file_from_url()}}}
}
\value{
data frame, or data table with argument \code{return_data_table = TRUE}
}
\description{
Load training summary statistics from differential analysis (DA)
that tests the effect of training on each feature within each sex.
Results are downloaded from a public Google Cloud Storage bucket.
}
\examples{
res = load_training_da("ACETYL", "HEART")
\dontrun{res = load_training_da("ATAC", "SKM-GN")}
res = load_training_da("IMMUNO", "ADRNL")
res = load_training_da("METAB", "SKM-GN", metareg = TRUE)
res = load_training_da("METAB", "SKM-GN", metareg = FALSE)
\dontrun{res = load_training_da("METHYL", "SKM-GN")}
res = load_training_da("PHOSPHO", "LIVER")
res = load_training_da("PROT", "KIDNEY")
res = load_training_da("TRNSCRPT", "BLOOD")
res = load_training_da("UBIQ", "HEART")
}
\seealso{
\link[MotrpacRatTraining6moData:TRAINING_DA]{MotrpacRatTraining6moData::TRAINING_DA}
}
