% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{load_uniprot_human_fasta}
\alias{load_uniprot_human_fasta}
\title{Load UniProt human canonical protein FASTA file}
\usage{
load_uniprot_human_fasta(scratchdir = ".")
}
\arguments{
\item{scratchdir}{character, directory in which the file from Google Cloud
Storage should be downloaded}
}
\value{
\link[Biostrings:XStringSet-class]{Biostrings::XStringSet} object returned from reading in
the FASTA file with \code{\link[Biostrings:XStringSet-io]{Biostrings::readAAStringSet()}}
}
\description{
Load UniProt human canonical protein FASTA file
}
\details{
The human proteome FATSA for canonical proteins was downloaded from UniProt on
2/3/2021 (UniProtKB query "reviewed:true AND proteome:up000005640").
}
\examples{
fasta = load_uniprot_human_fasta()
head(fasta)

}
\seealso{
\code{\link[=find_flanks]{find_flanks()}}
}
