% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment.R
\name{make_kegg_db}
\alias{make_kegg_db}
\title{Make KEGG database}
\usage{
make_kegg_db(kegg_db_destination)
}
\arguments{
\item{kegg_db_destination}{character, target directory for KEGG
database. Parent directories are created if they do not yet exist.}
}
\description{
Make KEGG database for \code{FELLA} pathway enrichment with metabolites.
Run internally in \code{\link[=cluster_pathway_enrichment]{cluster_pathway_enrichment()}}.
}
\seealso{
\code{\link[=cluster_pathway_enrichment]{cluster_pathway_enrichment()}}, \code{\link[=run_fella]{run_fella()}}
}
