% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_meta_regression.R
\name{merge_two_dea_dfs}
\alias{merge_two_dea_dfs}
\title{Concatenate data frames}
\usage{
merge_two_dea_dfs(x1, x2)
}
\arguments{
\item{x1}{data frame}

\item{x2}{data frame}
}
\description{
Concatenate (rbind) two data frames and expand the result to include all columns from both tables.
}
\examples{
merge_two_dea_dfs(
  data.frame(a=1:3,b=5:7),
  data.frame(a=1:3,b=5:7)
)
merge_two_dea_dfs(
  data.frame(a=1:3,b=5:7),
  data.frame(a=1:3,b=5:7,c=10:12)
)
merge_two_dea_dfs(
  data.frame(a=1:3,b=5:7,d=10:12),
  data.frame(a=1:3,b=5:7,c=10:12)
)
}
