% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_meta_analysis.R
\name{metab_meta_analysis}
\alias{metab_meta_analysis}
\title{Metabolomics timewise meta-analysis}
\usage{
metab_meta_analysis(tissue, input = NULL)
}
\arguments{
\item{tissue}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}

\item{input}{r data frame, custom input. To see the expected format, look at
a table returned by \code{\link[=load_metabolomics_da]{load_metabolomics_da()}} with \code{type="timewise"}.}
}
\value{
a named list, where \code{res$meta_analysis_res} is a list with one \code{\link[metafor:rma.uni]{metafor::rma.uni()}} object per
meta-analyzed feature, and \code{res$merged_res} is a data frame non-redundant measurements for all input features
with the following columns:
\describe{
\item{\code{feature_ID}}{character, MoTrPAC feature identifier}
\item{\code{tissue}}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}
\item{\code{dataset}}{character, metabolomics platform (metab-u-ionpneg, metab-u-lrpneg, metab-u-lrppos, metab-u-hilicpos, metab-u-rpneg, metab-u-rppos, metab-t-amines, metab-t-oxylipneg, metab-t-tca, metab-t-nuc, metab-t-acoa, metab-t-ka) the feature was measured in. 'meta-reg' specifies results from the metabolomics meta-regression for repeated features.}
\item{\code{assay}}{character, assay abbreviation, one of \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV} ("METAB")}
\item{\code{sex}}{character, one of 'male' or 'female'}
\item{\code{is_targeted}}{logical, is this a targeted platform?}
\item{\code{site}}{character, Chemical Analysis Site (CAS) name}
\item{\code{metabolite_refmet}}{character, RefMet name of metabolite}
\item{\code{logFC}}{double, log fold-change where the numerator is 'comparison_group', e.g., '1w', and the denominator is the group of sex-matched sedentary control animals}
\item{\code{logFC_se}}{double, standard error of the log fold-change}
\item{\code{zscore}}{numeric, z-score (either original t-score or z-score from meta-analysis)}
\item{\code{cv}}{double, feature coefficient of variation in the dataset}
\item{\code{comparison_group}}{character, time point of trained animals compared to the sex-matched sedentary control animals, one of '1w', '2w', '4w', '8w'}
\item{\code{p_value}}{double, unadjusted p-value for the difference between 'comparison_group' and the group of sex-matched sedentary control animals}
\item{\code{comparison_average_intensity}}{double, average intensity among the replicates in 'comparison_group'}
\item{\code{reference_average_intensity}}{double, average intensity among the replicates in the group of sex-matched sedentary control animals}
\item{\code{I2}}{double, heterogeneity coefficient.
Higher values mean the measurements between sites/platforms are more discordant.}
\item{\code{QEp}}{double, heterogeneity p-value. A smaller p-value means
the measurements between sites/platforms are more discordant.}
\item{\code{min_nominal_p}}{double, minimum nominal p-value for this sex
and time point across potentially multiple measurements in the input}
}
}
\description{
Perform meta-analysis for repeated measurements and return non-redundant results.
If there were no repeated measurements for a feature, the input summary statistics
are returned. \strong{These results were not used in the manuscript.}
}
\examples{
# Perform meta-analysis on gastrocnemius timewise summary statistics
res = metab_meta_analysis("SKM-GN")
}
