% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_meta_regression.R
\name{metab_meta_regression}
\alias{metab_meta_regression}
\title{Metabolomics meta-regression}
\usage{
metab_meta_regression(
  tissue,
  timewise_input = NULL,
  training_input = NULL,
  het_p_threshold = 0.001
)
}
\arguments{
\item{tissue}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}

\item{timewise_input}{r data frame, custom input. To see the expected format, look at
a table returned by \code{\link[=load_metab_da]{load_metab_da()}} with \code{type="timewise"}.}

\item{training_input}{r data frame, custom input. To see the expected format, look at
a table returned by \code{\link[=load_metab_da]{load_metab_da()}} with \code{type="training"}.}

\item{het_p_threshold}{numeric, meta-regression cases with a heterogeneity p-value
below this are considered to have high heterogeneity. Default: 0.001}
}
\value{
named list where \code{meta_reg_timewise_dea} is a data frame with
the adjusted timewise results, \code{training_meta_regression} is a data frame with the adjusted training
results, \code{meta_regression_results} is a named list with meta-regression
results per redundant metabolite, \code{meta_regression_models} is a table of the number of each
type of model used for meta-regression, and \code{metabolite_categories} is a named list
of the RefMet IDs of metabolites corresponding to each category described in the details.
}
\description{
Perform meta-regression for repeated measurements.
Return merged timewise and training differential analysis summary statistics,
where results for features with multiple measurements are replaced with the
meta-regression results when appropriate. This method was used to generate the
\link[MotrpacRatTraining6moData:METAB_DA_METAREG]{MotrpacRatTraining6moData::METAB_DA_METAREG} differential analysis results,
which are the version of results used in the manuscript analyses.
}
\details{
We try multiple models per repeated analyte:
\itemize{
\item{Model 1:}{Two random effects factors if \code{platform} and \code{is_targeted} are not redundant. Default optimization.}
\item{Model 2:}{Two random effects factors if \code{platform} and \code{is_targeted} are not redundant with alternative optimization.}
\item{Model 3:}{\code{platform} and \code{is_targeted} are redundant. Use a single RE factor with default optimization.
Also, use this if \code{QMp} is NA, which is an indication of over-parameterization of the model.}
\item{Model 4:}{\code{platform} and \code{is_targeted} are redundant. Use a single RE factor with alternative optimization.}
\item{Model 5:}{If all previous analyses failed, use a simple fixed-effects approach.}
}

After performing meta-regression, we define four categories of metabolites:
\enumerate{
\item measured once
\item measured multiple times, high heterogeneity, has a targeted platform
\item measured multiple times, high heterogeneity, no targeted platform
\item measured multiple times, low heterogeneity
}

For categories 1 and 3 we keep the results as is.
For category 2 we take the targeted data only.
Finally, for category 4 we take the meta-regression results.
}
\examples{
# Perform meta-regression for gastrocnemius
res = metab_meta_regression("SKM-GN")
names(res)
}
\seealso{
\code{\link[=load_metab_da]{load_metab_da()}}, \code{\link[=metabolite_meta_regression]{metabolite_meta_regression()}}, \link[MotrpacRatTraining6moData:METAB_DA_METAREG]{MotrpacRatTraining6moData::METAB_DA_METAREG}
}
