% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_differential_analysis.R
\name{metab_timewise_da}
\alias{metab_timewise_da}
\title{Metabolomics timewise differential analysis}
\usage{
metab_timewise_da(tissue)
}
\arguments{
\item{tissue}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}
}
\value{
a data frame with one row per metabolite, time, and sex combination:
\describe{
\item{\code{feature_ID}}{character, MoTrPAC feature identifier}
\item{\code{sex}}{character, one of 'male' or 'female'}
\item{\code{cv}}{double, coefficient of variation for case samples}
\item{\code{control_cv}}{double, coefficient of variation for control samples}
\item{\code{logFC}}{double, log fold-change where the numerator is 'comparison_group', e.g., '1w', and the denominator is the group of sex-matched sedentary control animals}
\item{\code{logFC_se}}{double, standard error of the log fold-change}
\item{\code{tscore}}{double, t statistic}
\item{\code{covariates}}{character, comma-separated list of adjustment variables or NA}
\item{\code{comparison_group}}{character, time point of trained animals compared to the sex-matched sedentary control animals, one of '1w', '2w', '4w', '8w'}
\item{\code{p_value}}{double, unadjusted p-value for the difference between 'comparison_group' and the group of sex-matched sedentary control animals}
\item{\code{comparison_average_intensity}}{double, average intensity among the replicates in 'comparison_group'}
\item{\code{reference_average_intensity}}{double, average intensity among the replicates in the group of sex-matched sedentary control animals}
\item{\code{dataset}}{character, the metabolomics platform in which the metabolite is detected}
}
}
\description{
Use \code{limma} to perform pairwise contrasts between each group of trained animals
and the sex-matched control group for a single tissue. Analysis is performed separately for males and
females.
}
\examples{
# Perform timewise differential analysis for metabolites in heart tissue
metab_timewise_da("HEART")

}
