% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_differential_analysis.R
\name{metab_training_da}
\alias{metab_training_da}
\title{Metabolomics training differential analysis}
\usage{
metab_training_da(tissue)
}
\arguments{
\item{tissue}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}
}
\value{
a data frame with one row per metabolite:
\describe{
\item{\code{feature_ID}}{character, MoTrPAC feature identifier}
\item{\code{dataset}}{character, the metabolomics platform in which the metabolite is detected}
\item{\code{groups_tested_female}}{character, timepoints used to perform the F-test in females.
Some tissues or assays are missing timepoints.}
\item{\code{groups_tested_male}}{character, timepoints used to perform the F-test in males.
Some tissues or assays are missing timepoints.}
\item{\code{fscore_male}}{double, F-statistic for males}
\item{\code{fscore_female}}{double, F-statistic for females}
\item{\code{p_value_male}}{double, nominal p-value for males}
\item{\code{p_value_female}}{double, nominal  p-value for females}
\item{\code{full_model}}{character, full model used in limma}
\item{\code{reduced_model}}{character, representation of the reduced model,
though not used by limma}
\item{\code{p_value}}{double, combined male and female nominal p-value using the sum of logs}
}
}
\description{
Use \code{limma} to test the effect of training across timepoints. The analysis
is performed separately for male and female rats and combined into a single
p-value using sum of logs.
}
\examples{
# Perform training differential analysis for metabolites in heart tissue
res = metab_training_da("HEART")

}
