% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_meta_regression.R
\name{metabolite_meta_regression}
\alias{metabolite_meta_regression}
\title{Meta-regression for a metabolite}
\usage{
metabolite_meta_regression(metabolite_refmet, timewise, training = NULL)
}
\arguments{
\item{metabolite_refmet}{character, RefMet name of metabolite}

\item{timewise}{data frame with all timewise summary statistics}

\item{training}{optional data frame with all training summary statistics}
}
\value{
a named list with meta-regression results, metabolite metadata, and
training summary statistics if \code{training} is not NULL.
}
\description{
Worker function wrapped by \code{\link[=metabolomics_meta_regression]{metabolomics_meta_regression()}}.
}
\details{
We try multiple models per analyte:
\itemize{
\item{Model 1:}{Two random effects factors if \code{platform} and \code{is_targeted} are not redundant. Default optimization.}
\item{Model 2:}{Two random effects factors if \code{platform} and \code{is_targeted} are not redundant with alternative optimization.}
\item{Model 3:}{\code{platform} and \code{is_targeted} are redundant. Use a single RE factor with default optimization.
Also, use this if \code{QMp} is NA, which is an indication of over-parameterization of the model.}
\item{Model 4:}{\code{platform} and \code{is_targeted} are redundant. Use a single RE factor with alternative optimization.}
\item{Model 5:}{If all previous analyses failed, use a simple fixed-effects approach.}
}

See \url{https://www.metafor-project.org/doku.php/tips:models_with_or_without_intercept} for some explanations.
See \url{https://www.publichealth.columbia.edu/research/population-health-methods/meta-regression} for general background,
including discussion on FE models for replicated experiments (as in our case).

"Use of a fixed effect meta-analysis model assumes all studies are estimating the same (common) treatment effect.":
\itemize{
\item Riley et al. \url{https://www-bmj-com.stanford.idm.oclc.org/content/342/bmj.d549}
}

Hypothetically, if all studies had an infinite sample size,
there would be no differences due to chance and the differences in study estimates would completely disappear.

Unlike the other assays, having greater sampling variance in
one of the sexes is not a problem for the model and we therefore do not need to split the sexes.

Most longitudinal meta-analyses include random effect terms per study. For examples see
\url{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0164898},
\url{https://metafor-project.org/doku.php/tips:two_stage_analysis}.
}
