% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_meta_regression.R
\name{metabolomics_meta_regression}
\alias{metabolomics_meta_regression}
\title{Metabolomics meta-regression}
\usage{
metabolomics_meta_regression(
  tissue,
  timewise_input = NULL,
  training_input = NULL,
  het_p_threshold = 0.001
)
}
\arguments{
\item{tissue}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}

\item{timewise_input}{r data frame, custom input. To see the expected format, look at
a table returned by \code{\link[=load_metabolomics_da]{load_metabolomics_da()}} with \code{type="timewise"}.}

\item{training_input}{r data frame, custom input. To see the expected format, look at
a table returned by \code{\link[=load_metabolomics_da]{load_metabolomics_da()}} with \code{type="training"}.}

\item{het_p_threshold}{numeric, meta-regression cases with a heterogeneity p-value
below this are considered to have high heterogeneity. Default: 0.001}
}
\value{
named list where \code{meta_reg_timewise_dea} is a data frame with
the adjusted timewise results, \code{training_meta_regression} is a data frame with the adjusted training
results, \code{meta_regression_results} is a named list with meta-regression
results per redundant metabolite, \code{meta_regression_models} is a table of the number of each
type of model used for meta-regression, and \code{metabolite_categories} is a named list
of the RefMet IDs of metabolites corresponding to each category described in the details.
}
\description{
Perform meta-regression for repeated measurements.
Return merged timewise and training differential analysis summary statistics,
where results for features with multiple measurements are replaced with the
meta-regression results when appropriate.
}
\details{
After performing meta-regression, we define four categories of metabolites:
\enumerate{
\item measured once
\item measured multiple times, high heterogeneity, has a targeted platform
\item measured multiple times, high heterogeneity, no targeted platform
\item measured multiple times, low heterogeneity
}

For categories 1 and 3 we keep the results as is.
For category 2 we take the targeted data only.
Finally, for category 4 we take the meta-regression results.
}
\examples{
# Perform meta-regression for gastrocnemius
res = metabolomics_meta_regression("SKM-GN")
names(res)
}
\seealso{
\code{\link[=load_metabolomics_da]{load_metabolomics_da()}}, \code{\link[=metabolite_meta_regression]{metabolite_meta_regression()}}
}
