% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotrpacRatTraining6mo-package.R
\docType{package}
\name{MotrpacRatTraining6mo-package}
\alias{MotrpacRatTraining6mo}
\alias{MotrpacRatTraining6mo-package}
\title{MotrpacRatTraining6mo: Analysis of the MoTrPAC endurance exercise training data in 6-month-old rats}
\description{
The goal of the Molecular Transducers of Physical Activity Consortium (MoTrPAC) is to study the molecular changes that occur during and after exercise and ultimately advance the understanding of how physical activity improves and preserves health. This package provides convenient analysis functions to explore the multi-tissue, multi-omic data generated by MoTrPAC's endurance exercise training study in 6-month-old rats.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MoTrPAC/MotrpacRatTraining6mo}
  \item \url{https://motrpac.github.io/MotrpacRatTraining6mo/}
}

}
\author{
\strong{Maintainer}: Nicole Gay \email{nicole.r.gay@gmail.com}

Authors:
\itemize{
  \item David Amar \email{ddam.am@gmail.com}
  \item Pierre Jean Beltran \email{pjeanbeltran@gmail.com}
  \item MoTrPAC Study Group [data contributor]
}

Other contributors:
\itemize{
  \item NIH Common Fund [funder]
  \item National Institute of Diabetes and Digestive and Kidney Diseases (NIDDK) [funder]
  \item National Institute of Arthritis and Musculoskeletal Diseases (NIAMS) [funder]
  \item National Institute on Aging (NIA) [funder]
  \item National Institute of Biomedical Imaging and Bioengineering [funder]
}

}
\keyword{internal}
