% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment.R
\name{pathway_hypergeom_test}
\alias{pathway_hypergeom_test}
\title{Custom pathway enrichment test}
\usage{
pathway_hypergeom_test(
  feature_to_gene,
  universe,
  cluster_res,
  pathway_member_list,
  source,
  gene_identifier_type,
  iterations,
  i,
  min_input_set_size,
  logfile,
  add_ensembl_intersection
)
}
\arguments{
\item{feature_to_gene}{data frame, map between \code{intersection_id_type} and gene symbols.
Columns must include "feature_ID", "gene_symbol", "ensembl_gene", and "kegg_id".
\link[MotrpacRatTraining6moData:FEATURE_TO_GENE]{MotrpacRatTraining6moData::FEATURE_TO_GENE} by default.}

\item{universe}{list of lists of character vectors,
named first by assay (i.e., \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV})
and then by tissue (i.e., \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}).
Vectors provide the full set of gene symbols associated with features tested
during differential analysis. For example, \code{universe$TRNSCRPT$LUNG}
should be a character vector of expressed genes in the lung, where the type
of gene identifier matches \code{gene_identifier_type}.
\code{[MotrpacRatTraining6moData::GENE_UNIVERSES]$gene_symbol} by default.}

\item{cluster_res}{Either a data frame or a list of lists.
If a data frame, it needs at least two columns: "feature" and "cluster".
The "feature" column should be in the format
'\link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV};\link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV};feature_ID'.
If a list of lists, each sublist must be named with the cluster name (character string),
and the values must be features in the format
'\link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV};\link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV};feature_ID'.}

\item{pathway_member_list}{named list of character vectors where names are pathway names
and values are pathway members. Pathway members must match values in the
\code{gene_identifer_type} column of \code{feature_to_gene}.}

\item{source}{optional character string to define the source of \code{pathway_member_list}}

\item{gene_identifier_type}{character, column in \code{feature_to_gene} that
matches the gene identifier type in \code{universe}.
"gene_symbol" by default.}

\item{iterations}{data frame passed from \code{\link[=custom_cluster_pathway_enrichment]{custom_cluster_pathway_enrichment()}}
that defines each iteration of this function}

\item{i}{integer index passed from \code{\link[=custom_cluster_pathway_enrichment]{custom_cluster_pathway_enrichment()}} that specifies
which row of \code{iterations} to use}

\item{min_input_set_size}{integer, input must have this minimum number of
unique mappable gene IDs to attempt enrichment}

\item{logfile}{optional character, path to log of failed iterations}

\item{add_ensembl_intersection}{bool, whether to add a \code{intersection_ensembl}
column, which converts gene IDs in the intersection to Ensembl IDs}
}
\value{
data table with enrichment results, or NULL if no enrichment results were returned:
\describe{
\item{\code{term_size}}{integer, number of genes that are annotated to the term}
\item{\code{query_size}}{integer, number of genes that were included in the query}
\item{\code{intersection_size}}{integer, the number of genes in the input query that are annotated to the corresponding term}
\item{\code{term_id}}{character, unique term/pathway identifier}
\item{\code{source}}{character, the abbreviation of the data source for the term/pathway}
\item{\code{term_name}}{character, term/pathway name}
\item{\code{effective_domain_size}}{integer, the total number of genes in the universe used for the hypergeometric test}
\item{\code{intersection}}{character, input gene IDs that intersect with the term/pathway}
\item{\code{computed_p_value}}{double, nominal hypergeometric p-value}
\item{\code{cluster}}{character, cluster label}
\item{\code{tissue}}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}
\item{\code{ome}}{character, assay abbreviation, one of \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV}}
}
}
\description{
Worker function run by \code{\link[=custom_cluster_pathway_enrichment]{custom_cluster_pathway_enrichment()}}.
Not intended to be run independently.
}
\seealso{
\code{\link[=custom_cluster_pathway_enrichment]{custom_cluster_pathway_enrichment()}}
}
