% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_graphical_clustering.R
\name{plot_features_per_cluster}
\alias{plot_features_per_cluster}
\title{Plot feature composition of clusters}
\usage{
plot_features_per_cluster(cluster_res)
}
\arguments{
\item{cluster_res}{Either a data frame or a list of lists.
If a data frame, it needs at least two columns: "feature" and "cluster".
The "feature" column should be in the format
'\link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV};\link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV};feature_ID'.
If a list of lists, each sublist must be named with the cluster name (character string),
and the values must be features in the format
'\link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV};\link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV};feature_ID'.}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot number of features in each cluster and fractions of features that belong to each ome and tissue.
}
\examples{
# Get top 10 largest paths, nodes, edges in gastrocnemius 
# Exclude additional 8-week nodes
clusters = extract_tissue_sets("SKM-GN", k=10, add_week8=FALSE)
# Select paths only 
clusters = clusters[grepl("->", names(clusters))]

# Plot distribution of features
plot_features_per_cluster(clusters)

}
