% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_outliers.R
\name{plot_pcs}
\alias{plot_pcs}
\title{Plot 2D scatter plot of principal components}
\usage{
plot_pcs(pcA, pcB, pcax, outliers, pca, title = NULL)
}
\arguments{
\item{pcax}{e.g., \code{prcomp(data)$x}}

\item{outliers}{list of viallabels corresponding to PC outliers}

\item{pca}{result returned by \code{prcomp()}}

\item{title}{optional plot title}

\item{pcaA}{string of PC for x-axis, e.g. "PC1"; also a column name in \code{pcax}}

\item{pcaB}{string of PC for y-axis, e.g. "PC2"; also a column name in \code{pcax}}
}
\value{
ggplot object
}
\description{
Internal function to plot pairs of principal components and highlight indicated outliers.
Each point is a sample.
}
\examples{
TODO

}
\seealso{
[call_pca_outliers()]
}
