% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_datasets.R
\name{prep_data_TRNSCRPT}
\alias{prep_data_TRNSCRPT}
\title{Title}
\usage{
prep_data_TRNSCRPT(tissue, sex, outliers = NULL)
}
\arguments{
\item{tissue}{character, tissue abbreviation, one of [TISSUE_ABBREV], excluding "PLASMA"}

\item{sex}{character, "male", "female", or "all"}

\item{outliers}{optional, vector of vial labels to exclude}
}
\value{

}
\description{
Description
}
\examples{
TODO

}
