% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{prepare_gsea_input}
\alias{prepare_gsea_input}
\title{Prepare GCT file for GSEA}
\usage{
prepare_gsea_input(
  tissue = NULL,
  assay = NULL,
  outdir = ".",
  outfile_prefix = NULL,
  input = NULL,
  cast_vars = c("sex", "comparison_group")
)
}
\arguments{
\item{tissue}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}

\item{assay}{character, assay abbreviation, one of "PROT", "PHOSPHO", "TRNSCRPT", "ACETYL", "UBIQ"}

\item{outdir}{character, output directory for GCT file. The directory is created if it does not already exist.
Current directory by default.}

\item{outfile_prefix}{character, prefix for output GCT file. By default, this prefix
includes the specified tissue and assay and current date. Must be specified for custom input data.}

\item{input}{optional data frame if the user wants to perform this analysis for
a custom set of results. Required columns are "tscore", "gene_symbol", and \code{cast_vars}.}

\item{cast_vars}{character vector of column names in the differential analysis results
that are used to convert the table from long to wide format, with t-scores as the value variable.
See \code{\link[data.table:dcast.data.table]{data.table::dcast()}} for more details. Default: "sex", "comparison_group"}
}
\value{
character, path of the GCT file
}
\description{
Take a set of differential analysis results, either for a specific tissue and ome or
from a user-supplied table, convert to GCT format, and write to file for GSEA.
}
\details{
T-scores from the timewise differential analysis results are used for scores.
Feature-level data is summarized into gene-level data using the maximum absolute t-score.
}
\examples{
prepare_gsea_input("HEART","PROT",outdir="/tmp")
prepare_gsea_input("LIVER","PHOSPHO",outdir="/tmp")
prepare_gsea_input("LIVER","ACETYL",outdir="/tmp")
prepare_gsea_input("LIVER","UBIQ",outdir="/tmp")
prepare_gsea_input("LIVER","TRNSCRPT",outdir="/tmp")

# "custom" input
res = combine_da_results(tissues = "KIDNEY", assays = "PROT")
# add dummy column
res$gene_symbol = res$feature_ID
prepare_gsea_input(input=res, outdir="/tmp", outfile_prefix="KIDNEY_PROT")

}
