% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{prepare_ptmsea_input}
\alias{prepare_ptmsea_input}
\title{Prepare PTM-SEA input}
\usage{
prepare_ptmsea_input(
  tissue = NULL,
  fasta = NULL,
  rat_to_human_ptm_map = MotrpacRatTraining6moData::RAT_TO_HUMAN_PHOSPHO,
  outdir = ".",
  outfile_prefix = NULL,
  input = NULL,
  cast_vars = c("sex", "comparison_group")
)
}
\arguments{
\item{tissue}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}. Must be specified if cusotm \code{input} is not provided.}

\item{fasta}{\link[Biostrings:XStringSet-class]{Biostrings::XStringSet} object returned from reading in
a human protein FASTA file with \code{\link[Biostrings:XStringSet-io]{Biostrings::readAAStringSet()}}. Names of the
\link[Biostrings:XStringSet-class]{Biostrings::XStringSet} object should be set to the human protein accession,
e.g., "Q96QG7". If not specified, the result of \code{\link[=load_uniprot_human_fasta]{load_uniprot_human_fasta()}} is used, which
returns the version of the UniProt human protein FASTA used in the manuscript.}

\item{rat_to_human_ptm_map}{data frame with columns "ptm_id_rat_refseq" and "ptm_id_human_uniprot"
used to map PTMs from rat to human. Default: \link[MotrpacRatTraining6moData:RAT_TO_HUMAN_PHOSPHO]{MotrpacRatTraining6moData::RAT_TO_HUMAN_PHOSPHO}}

\item{outdir}{character, output directory for GCT file. The directory is created if it does not already exist.
Current directory by default.}

\item{outfile_prefix}{character, prefix for output GCT file. By default, this prefix
includes the specified tissue and current date. Must be specified for custom input data.}

\item{input}{optional data frame if the user wants to perform this analysis for
a custom set of differential analysis results. Required columns are "feature_ID", "tscore", and \code{cast_vars}.}

\item{cast_vars}{character vector of column names in the differential analysis results
that are used to convert the table from long to wide format, with t-scores as the value variable.
See \code{\link[tidyr:pivot_wider]{tidyr::pivot_wider()}} for more details. Default: "sex", "comparison_group"}
}
\value{
character, full path to GCT file
}
\description{
Write a GCT file with timewise differential analysis t-scores as values and
human flanking sequences as IDs. \code{rat_to_human_ptm_map} is used to map from rat
to human phosphosites, and a human protein FASTA file is used to add human
flanking sequences.
}
\examples{
\dontrun{
# Using existing differential analysis results
prepare_ptmsea_input("HEART", outdir="/tmp")

# Using a "custom" input
res = combine_da_results(tissues = "HEART", assays = "PHOSPHO")
# add dummy column
prepare_ptmsea_input(input=res, outdir="/tmp", outfile_prefix="HEART_PHOSPHO")
}
}
\seealso{
\code{\link[=load_uniprot_human_fasta]{load_uniprot_human_fasta()}}, \code{\link[=find_flanks]{find_flanks()}}, \link[MotrpacRatTraining6moData:RAT_TO_HUMAN_PHOSPHO]{MotrpacRatTraining6moData::RAT_TO_HUMAN_PHOSPHO}
}
