% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_datasets.R
\name{preprocess_pass1b_atacseq_gcp}
\alias{preprocess_pass1b_atacseq_gcp}
\title{Prepare ATAC-seq dataset}
\usage{
preprocess_pass1b_atacseq_gcp(
  TISSUE_CODE,
  outliers = NULL,
  gsutil_path = "~/google-cloud-sdk/bin/gsutil",
  scratch = "/oak/stanford/groups/smontgom/nicolerg/tmp",
  parallel = T
)
}
\arguments{
\item{TISSUE_CODE}{}

\item{outliers}{}

\item{gsutil_path}{}

\item{scratch}{}

\item{parallel}{}
}
\value{

}
\description{
Retrieve and format ATAC-seq sample-level data and metadata for a given tissue.
}
\examples{
TODO

}
