% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_datasets.R
\name{preprocess_pass1b_rnaseq_gcp}
\alias{preprocess_pass1b_rnaseq_gcp}
\title{Title}
\usage{
preprocess_pass1b_rnaseq_gcp(
  TISSUE_CODE,
  SEX,
  gsutil_path = "~/google-cloud-sdk/bin/gsutil",
  scratch = "/oak/stanford/groups/smontgom/nicolerg/tmp",
  outliers = NULL,
  parallel = F
)
}
\arguments{
\item{TISSUE_CODE}{}

\item{SEX}{}

\item{gsutil_path}{}

\item{scratch}{}

\item{outliers}{}

\item{parallel}{}
}
\value{

}
\description{
Title
}
\examples{
TODO

}
