% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_differential_analysis.R
\name{proteomics_timewise_da}
\alias{proteomics_timewise_da}
\title{Proteomics timewise differential analysis}
\usage{
proteomics_timewise_da(assay, tissue, exclude_outliers = TRUE)
}
\arguments{
\item{assay}{character, abbreviation for proteomics assay to be analyzed as defined by \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV}.
One of the following: PROT, PHOSPHO, ACETYL, UBIQ}

\item{tissue}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}

\item{exclude_outliers}{bool, whether to remove sample outliers specified in
\link[MotrpacRatTraining6moData:OUTLIERS]{MotrpacRatTraining6moData::OUTLIERS}. \code{TRUE} by default.}
}
\value{
a data frame with one row per proteomics feature per contrast (usually 8 rows per feature):
\describe{
\item{\code{feature_ID}}{character, MoTrPAC feature identifier}
\item{\code{sex}}{character, one of 'male' or 'female'}
\item{\code{comparison_group}}{character, time point of trained animals compared to the sex-matched sedentary control animals, one of '1w', '2w', '4w', '8w'}
\item{\code{assay}}{character, assay abbreviation, one of \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV}}
\item{\code{assay_code}}{character, assay code used in data release. See \link[MotrpacBicQC:assay_codes]{MotrpacBicQC::assay_codes}.}
\item{\code{tissue}}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}
\item{\code{tissue_code}}{character, tissue code used in data release. See \link[MotrpacBicQC:bic_animal_tissue_code]{MotrpacBicQC::bic_animal_tissue_code}.}
\item{\code{removed_samples}}{character, comma-separated list of outliers (vial labels) removed from differential analysis}
\item{\code{logFC}}{double, log fold-change where the numerator is 'comparison_group', e.g., '1w', and the denominator is the group of sex-matched sedentary control animals}
\item{\code{logFC_se}}{double, standard error of the log fold-change}
\item{\code{tscore}}{double, t statistic}
\item{\code{p_value}}{double, unadjusted p-value for the difference between 'comparison_group' and the group of sex-matched sedentary control animals}
\item{\code{comparison_average_intensity}}{double, average intensity among the replicates in 'comparison_group'}
\item{\code{reference_average_intensity}}{double, average intensity among the replicates in the group of sex-matched sedentary control animals}
\item{\code{numNAs}}{integer, number of missing values in the current 'comparison_group' and sex-matched sedentary control animals}
}
}
\description{
Timewise differential analysis for the proteome, phosphoproteome, acetylome, and ubiquitylome.
Use limma to perform pairwise contrasts between each group of trained animals
and the sex-matched control group for a single tissue and proteomics assay.
Analysis is performed separately for males and females.
}
\examples{
# Run timewise differential analysis for heart proteins
proteomics_timewise_da("PROT","HEART")
}
