% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_differential_analysis.R
\name{proteomics_training_da}
\alias{proteomics_training_da}
\title{Proteomics training differential analysis}
\usage{
proteomics_training_da(assay, tissue, exclude_outliers = TRUE)
}
\arguments{
\item{assay}{character, abbreviation for proteomics assay to be analyzed as defined by \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV}.
One of the following: PROT, PHOSPHO, ACETYL, UBIQ}

\item{tissue}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}

\item{exclude_outliers}{bool, whether to remove sample outliers specified in
\link[MotrpacRatTraining6moData:OUTLIERS]{MotrpacRatTraining6moData::OUTLIERS}. \code{TRUE} by default.}
}
\value{
a data frame with one row per proteomics feature:
\describe{
\item{\code{feature_ID}}{character, MoTrPAC feature identifier}
\item{\code{assay}}{character, assay abbreviation, one of \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV}}
\item{\code{assay_code}}{character, assay code used in data release. See \link[MotrpacBicQC:assay_codes]{MotrpacBicQC::assay_codes}.}
\item{\code{tissue}}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}
\item{\code{tissue_code}}{character, tissue code used in data release. See \link[MotrpacBicQC:bic_animal_tissue_code]{MotrpacBicQC::bic_animal_tissue_code}.}
\item{\code{removed_samples}}{character, comma-separated list of outliers (vial labels) removed from differential analysis}
\item{\code{fscore_male}}{double, F statistic for males}
\item{\code{fscore_female}}{double, F statistic for females}
\item{\code{p_value_male}}{double, nominal F-test p-value for males}
\item{\code{p_value_female}}{double, nominal F-test p-value for females}
\item{\code{full_model}}{character, full model used in F-test}
\item{\code{reduced_model}}{character, reduced model used in F-test}
\item{\code{p_value}}{double, combined male and female nominal p-value using the sum of logs}
}
}
\description{
Training differential analysis for the proteome, phosphoproteome, acetylome, and ubiquitylome.
Use limma to perform an F-test to test the effect of training
across time points. Analysis is performed separately for males and females.
}
\examples{
# Run training differential analysis for heart proteins
res = proteomics_training_da("PROT","HEART")
}
