% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_differential_analysis.R
\name{proteomics_training_dea}
\alias{proteomics_training_dea}
\title{Title}
\usage{
proteomics_training_dea(assay_abbrev, tissue_abbrev)
}
\arguments{
\item{assay_abbrev}{Abbreviation for proteomics assay to be analyzed as defined by ASSAY_ABBREV.
One of the following: PROT, PHOSPHO, ACETYL, UBIQ}

\item{tissue_abbrev}{Abbreviation for proteomics tissue to be analyzed as defined by TISSUE_ABBREV.
One of the following: PROT, PHOSPHO, ACETYL, UBIQ}
}
\value{
A data frame containing differential enrichment results
}
\description{
Title
}
\examples{
proteomics_training_dea("PROT","HEART")
}
