% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment_networks.R
\name{replace_ensembl_with_symbol}
\alias{replace_ensembl_with_symbol}
\title{Replace Ensembl ID with gene symbol}
\usage{
replace_ensembl_with_symbol(x, map, return_N = TRUE, collapse = TRUE)
}
\arguments{
\item{x}{character, comma-separated Ensembl IDs}

\item{map}{data table, mapping between Ensembl IDs and gene symbols.
Must include the columns "ensembl_gene" and "gene_symbol".}

\item{return_N}{boolean, whether to prepend the concatenated
gene symbols with the number of unique gene symbols.
\code{TRUE} by default.}

\item{collapse}{boolean, whether to collapse the gene symbols into a
comma-separated string. If not, return a vector of unique gene symbols.
\code{TRUE} by default.}
}
\value{
either a string of comma-separated gene symbols,
a vector of unique gene symbols,
or a named list with two values ("N" and "genes")
depending on the values of \code{return_N} and \code{collapse}.
}
\description{
Function is used internally in \code{\link[=enrichment_network_vis]{enrichment_network_vis()}}.
}
\seealso{
\code{\link[=enrichment_network_vis]{enrichment_network_vis()}}
}
\keyword{internal}
