% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrbs_pipeline.R
\name{rrbs_differential_analysis}
\alias{rrbs_differential_analysis}
\title{Differential analysis wrapper for RRBS data}
\usage{
rrbs_differential_analysis(
  y,
  PHENO,
  METHYL_META,
  verbose = T,
  adj_pct_unaligned = F,
  samples_to_remove = NULL,
  edger_tol = 1e-05,
  dataset_name = ""
)
}
\arguments{
\item{PHENO}{A data frame with a row per sample. Can be retreived using data(PHENO).
Contains at least the following columns: sex, group}

\item{METHYL_META}{A data frame with a row per sample. Can be retreived using data(METHYL_META).
Contains the RRBS pipeline QA/QC scores. Contains at least the following columns: pct_Unaligned.}

\item{verbose}{A logical. TRUE: comments about the pipeline progress are printed.}

\item{adj_pct_unaligned}{A logical. TRUE: adjust for percent unaligned reads. Default is FALSE.}

\item{samples_to_remove}{A character vector. Contains the ids of the samples that should be removed
(e.g., identified outliers or failed samples).}

\item{edger_tol}{A number. An internal parameter of edgeR. Default is 1e-05. Consider increasing if the algorithm takes too long.}

\item{dataset_name}{A character. The name of the current dataset. Will be added to the output.}

\item{yall}{A DGEList object. 
yall$genes is a metadata data frame with the locus coordinates (see details), and these fields at minimum:
Chr, EntrezID, Symbol, and Strand}
}
\value{
A list.First item is called timewise and will contain the contrast-specific differential analysis results.
Second item is called training and contains the overall training-level significance per locus.
}
\description{
Differential analysis wrapper for RRBS data
}
\examples{
library(MotrpacRatTraining6moData)
data(PHENO)
data(METHYL_META)

# Raw data in RData file is available at: 
https://drive.google.com/drive/folders/1_vkqPc8uULIiCTHDW8nuNHZW4jzzLHxy?usp=sharing
yall = get(load("motrpac_pass1b-06_t55-gastrocnemius_epigen-rrbs_bismark-cov.RData"))

# Alternatively, you can use the processed datasets
(TBD)

# for the simplicity of this example, we subset the data to 5000 loci
y = yall[1:5000,]
dea_res = rrbs_differential_analysis(y,PHENO,METHYL_META,adj_pct_unaligned=T)
head(dea_res$timewise)
head(dea_res$training)
}
