% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment.R
\name{run_single_enrichment}
\alias{run_single_enrichment}
\title{Run single iteration of pathway enrichment}
\usage{
run_single_enrichment(
  feature_to_gene_map,
  universe,
  cluster_res,
  gene_identifier_type,
  databases,
  iterations,
  i,
  min_input_set_size,
  fella.data,
  fella_method,
  logfile = "/dev/null",
  maxattempt = 50
)
}
\arguments{
\item{feature_to_gene_map}{data table, map between \code{intersection_id_type} and gene symbols.
Columns must include "feature_ID", "gene_symbol", "ensembl_gene", and "kegg_id".}

\item{universe}{character vector of genes in the universe or background}

\item{cluster_res}{data frame output by \code{\link[=check_cluster_res_format]{check_cluster_res_format()}}}

\item{gene_identifier_type}{column in \code{feature_to_gene_map} that
matches the gene identifier type in \code{universe}}

\item{databases}{character vector of g:Profiler pathway databases to query.
"KEGG" and "REAC" (REACTOME) by default.
Current options include: GO (GO:BP, GO:MF, GO:CC to select a particular GO branch),
KEGG, REAC, TF, MIRNA, CORUM, HP, HPA, WP.
See \href{https://rdrr.io/cran/gprofiler2/man/gost.html}{gprofiler2 documentation} for an up-to-date list.}

\item{iterations}{data frame passed from \code{\link[=cluster_pathway_enrichment]{cluster_pathway_enrichment()}}
that defines each iteration of this function}

\item{i}{integer index passed from \code{\link[=cluster_pathway_enrichment]{cluster_pathway_enrichment()}} that specifies
which row of \code{iterations} to use}

\item{min_input_set_size}{integer, input must have this minimum number of
unique mappable gene IDs to attempt enrichment with \code{\link[gprofiler2:gost]{gprofiler2::gost()}}}

\item{fella.data}{FELLA database returned by \code{\link[FELLA:data-funs]{FELLA::loadKEGGdata()}} passed to \code{\link[=run_fella]{run_fella()}}}

\item{fella_method}{character, enrichment method for \code{\link[FELLA:enrich-funs]{FELLA::enrich()}},
one of "hypergeom" or "diffusion", passed to \code{\link[=run_fella]{run_fella()}}}

\item{logfile}{optional character, path to log of failed iterations}

\item{maxattempt}{integer, max number of consecutive null results from \code{\link[gprofiler2:gost]{gprofiler2::gost()}}
before giving up}
}
\value{
data table with enrichment results, or NULL if no enrichment results were returned
}
\description{
Run single iteration of pathway enrichment (FELLA for METAB or gprofiler for all other omes)
NOT intended to be run directly. Run within \code{\link[=cluster_pathway_enrichment]{cluster_pathway_enrichment()}}.
}
\seealso{
\code{\link[=cluster_pathway_enrichment]{cluster_pathway_enrichment()}}, \code{\link[=run_fella]{run_fella()}}
}
\keyword{internal}
