% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{STOPWORDS}
\alias{STOPWORDS}
\title{Stop words}
\format{
Unnamed character vector
}
\usage{
STOPWORDS
}
\description{
Stop words from \code{\link[tm:stopwords]{tm::stopwords()}}. Used in \code{\link[=cleanup]{cleanup()}}.
}
\keyword{datasets}
