% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_data_set.R
\name{tissue_abbreviations}
\alias{tissue_abbreviations}
\title{Return list of accepted tissue abbreviations}
\usage{
tissue_abbreviations(alphabetical_order = TRUE)
}
\arguments{
\item{alphabetical_order}{bool, whether to alphabetize the tissue abbreviations.
\code{TRUE} by default. If \code{FALSE}, return them in the biological order they
appear in manuscript figures.}
}
\value{
a vector of strings
}
\description{
Return tissue abbreviations for the 20 tissues analyzed in the manuscript.
}
\examples{
tissue_abbreviations()
tissue_abbreviations(alphabetical_order = FALSE)

}
