% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_outliers.R
\name{transcript_call_outliers}
\alias{transcript_call_outliers}
\title{Call RNA-seq sample outliers}
\usage{
transcript_call_outliers(tissues)
}
\arguments{
\item{tissues}{character vector of tissue abbreviations for which to call RNA-seq outliers.
See \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV} for accepted values.}
}
\value{
NULL if there are no outliers, or a data frame with three columns and one row per outlier:
\describe{
\item{\code{viallabel}}{character, sample identifier}
\item{\code{tissue}}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}
\item{\code{reason}}{character, PC(s) in which the sample was flagged}
}
}
\description{
Identify samples that are outside of 5 times the interquartile range of principal
components that explain at least 5\% of variance in each tissue. Use only the 1000 most variable genes.
This specifies RNA-seq outliers excluded from differential analysis by MoTrPAC.
}
\examples{
transcript_call_outliers("SKM-VL")
\dontrun{
transcript_call_outliers(c("SKM-GN","BLOOD"))
}
}
\seealso{
\code{\link[=call_pca_outliers]{call_pca_outliers()}} for workhorse function and \code{\link[=plot_pcs]{plot_pcs()}} for plotting function
}
