% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data.R
\name{transcript_normalize_counts}
\alias{transcript_normalize_counts}
\title{Filter and normalize raw RNA-seq counts}
\usage{
transcript_normalize_counts(
  tissue,
  min_cpm = 0.5,
  min_num_samples = 2,
  norm_method = "TMM"
)
}
\arguments{
\item{tissue}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}

\item{min_cpm}{double, retain genes with more than \code{min_cpm} counts per million in at least \code{min_num_samples} samples}

\item{min_num_samples}{double, retain genes with more than \code{min_cpm} counts per million in at least \code{min_num_samples} samples}

\item{norm_method}{character, one of \code{c("TMM","TMMwsp","RLE","upperquartile","none")}. "TMM" by default.}
}
\value{
data frame where row names are feature_ID and column names are viallabel
}
\description{
Get raw counts and return the filtered and normalized data.
}
\examples{
norm_data = transcript_normalize_counts("LUNG")

}
\seealso{
\link[MotrpacRatTraining6moData:TRNSCRPT_NORM_DATA]{MotrpacRatTraining6moData::TRNSCRPT_NORM_DATA}
}
