% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_datasets.R
\name{transcript_prep_data}
\alias{transcript_prep_data}
\title{Preprocess RNA-seq data}
\usage{
transcript_prep_data(
  tissue,
  sex = NULL,
  covariates = c("pct_globin", "RIN", "pct_umi_dup", "median_5_3_bias"),
  outliers = OUTLIERS$viallabel[OUTLIERS$assay == "TRNSCRPT"]
)
}
\arguments{
\item{tissue}{character, tissue abbreviation, one of \link{TISSUE_ABBREV}}

\item{sex}{NULL to return data from both sexes or one of "male" or "female" to return data from the specified sex}

\item{covariates}{character vector of covariates that correspond to column names of \link{TRNSCRPT_META}.
Defaults to covariates that were used for the manuscript.}

\item{outliers}{vector of viallabels to exclude during differential analysis. Defaults
to \code{OUTLIERS$viallabel[OUTLIERS$assay == "TRNSCRPT"]}}
}
\value{
named list of five items:
\describe{
\item{\code{metadata}}{TODO}
\item{\code{covariates}}TODO
\item{\code{filt_counts}}{TODO}
\item{\code{norm_data}}{TODO}
\item{\code{outliers}}{TODO}
}
}
\description{
Preprocess RNA-seq data
}
\examples{
gastroc_data = transcript_prep_data("SKM-GN")
gastroc_data = transcript_prep_data("SKM-GN", outliers = NULL)
gastroc_data = transcript_prep_data("SKM-GN", covariates = NULL, outliers = NULL)
gastroc_data = transcript_prep_data("SKM-GN", covariates = NULL, outliers = NULL, sex = "male")

}
\seealso{
\link{OUTLIERS}
}
