% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_differential_analysis.R
\name{transcript_timewise_da}
\alias{transcript_timewise_da}
\title{RNA-seq timewise differential analysis}
\usage{
transcript_timewise_da(
  tissue,
  covariates = c("pct_globin", "RIN", "pct_umi_dup", "median_5_3_bias"),
  outliers = na.omit(MotrpacRatTraining6moData::OUTLIERS$viallabel),
  add_shrunk_logfc = TRUE,
  rdata_outfile = NULL,
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{tissue}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}

\item{covariates}{character vector of covariates that correspond to column names of \link[MotrpacRatTraining6moData:TRNSCRPT_META]{MotrpacRatTraining6moData::TRNSCRPT_META}.
Defaults to covariates that were used for the manuscript.}

\item{outliers}{vector of viallabels to exclude during differential analysis. Defaults
to \code{[MotrpacRatTraining6moData::OUTLIERS\$viallabel[[MotrpacRatTraining6moData::OUTLIERS]$assay == "TRNSCRPT"]}}

\item{add_shrunk_logfc}{boolean, whether to calculate shrunk log fold-changes in addition to standard log fold-changes}

\item{rdata_outfile}{NULL or path in which to save DESeq2 objects in an RData file}

\item{overwrite}{boolean, whether to overwrite the file if \code{rdata_outfile} exists}

\item{verbose}{boolean, whether to print messages}
}
\value{
a data frame with one row per gene per contrast (usually 8 rows per gene):
\describe{
\item{\code{feature_ID}}{character, MoTrPAC feature identifier}
\item{\code{sex}}{character, one of 'male' or 'female'}
\item{\code{comparison_group}}{character, time point of trained animals compared to the sex-matched sedentary control animals, one of '1w', '2w', '4w', '8w'}
\item{\code{assay}}{character, assay abbreviation, one of \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV}}
\item{\code{assay_code}}{character, assay code used in data release. See \link[MotrpacBicQC:assay_codes]{MotrpacBicQC::assay_codes}.}
\item{\code{tissue}}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}
\item{\code{tissue_code}}{character, tissue code used in data release. See \link[MotrpacBicQC:bic_animal_tissue_code]{MotrpacBicQC::bic_animal_tissue_code}.}
\item{\code{covariates}}{character, comma-separated list of adjustment variables}
\item{\code{removed_samples}}{character, comma-separated list of outliers (vial labels) removed from differential analysis}
\item{\code{logFC}}{double, log fold-change where the numerator is 'comparison_group', e.g., '1w', and the denominator is the group of sex-matched sedentary control animals}
\item{\code{logFC_se}}{double, standard error of the log fold-change}
\item{\code{shrunk_logFC}}{double, log fold-change shrunk with \code{type = 'ashr'} and \code{optmethod = 'mixSQP'},
only if \code{add_shrunk_logfc = TRUE}}
\item{\code{shrunk_logFC_se}}{double, standard error of \code{shrunk_logFC}, only if \code{add_shrunk_logfc = TRUE}}
\item{\code{zscore}}{double, Wald statistic}
\item{\code{p_value}}{double, unadjusted p-value for the difference between 'comparison_group' and the group of sex-matched sedentary control animals}
\item{\code{comparison_average_intensity}}{double, average intensity among the replicates in 'comparison_group'}
\item{\code{comparison_average_intensity_se}}{double, standard error of 'comparison_average_intensity'}
\item{\code{reference_average_intensity}}{double, average intensity among the replicates in the group of sex-matched sedentary control animals}
\item{\code{reference_average_intensity_se}}{double, standard error of 'reference_average_intensity'}
}
}
\description{
Use DESeq2 to perform pairwise contrasts between each group of trained animals
and the sex-matched control group for a single tissue. Analysis is performed separately for males and
females.
}
\examples{
\dontrun{
# Perform differential analysis for expressed genes in brown adipose tissue 
# with default parameters, i.e., outliers and covariates used for the manuscript; 
# calculate both standard and shrunk log fold-changes
da = transcript_timewise_da("BAT")

# Same as above but don't calculate shrunk log fold-changes
da = transcript_timewise_da("BAT", add_shrunk_logfc = FALSE)

# Same as the first example but save the [DESeq2::DESeq2()] DESeqResults objects in an RData file 
da = transcript_timewise_da("BAT", rdata_outfile = "~/test/BAT_RNA_DA.RData", overwrite = TRUE)
}
}
