% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_differential_analysis.R
\name{transcript_timewise_dea}
\alias{transcript_timewise_dea}
\title{RNA-seq timewise differential analysis}
\usage{
transcript_timewise_dea(
  tissue,
  covariates = c("pct_globin", "RIN", "pct_umi_dup", "median_5_3_bias"),
 
    outliers = na.omit(MotrpacRatTraining6moData::OUTLIERS$viallabel[MotrpacRatTraining6moData::OUTLIERS$assay
    == "TRNSCRPT"]),
  add_shrunk_logfc = TRUE,
  rdata_outfile = NULL,
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{tissue}{@eval tissue()}

\item{covariates}{character vector of covariates that correspond to column names of \link[MotrpacRatTraining6moData:TRNSCRPT_META]{MotrpacRatTraining6moData::TRNSCRPT_META}.
Defaults to covariates that were used for the manuscript.}

\item{outliers}{vector of viallabels to exclude during differential analysis. Defaults
to \code{[MotrpacRatTraining6moData::OUTLIERS\$viallabel[[MotrpacRatTraining6moData::OUTLIERS]$assay == "TRNSCRPT"]}}

\item{add_shrunk_logfc}{boolean, whether to calculate shrunk log fold-changes in addition to standard log fold-changes}

\item{rdata_outfile}{NULL or path in which to save DESeq2 objects in an RData file}

\item{overwrite}{boolean, whether to overwrite the file if \code{rdata_outfile} exists}

\item{verbose}{boolean, whether to print the DESeq2 design string}
}
\value{
a data frame with one row per gene per contrast (usually 8 rows per gene):
\describe{
\item{\code{feature_ID}}{@eval feature_ID()}
\item{\code{sex}}{@eval sex()}
\item{\code{comparison_group}}{@eval comparison_group()}
\item{\code{assay}}{@eval assay()}
\item{\code{assay_code}}{@eval assay_code()}
\item{\code{tissue}}{@eval tissue()}
\item{\code{tissue_code}}{@eval tissue_code()}
\item{\code{covariates}}{comma-separated list of adjustment variables}
\item{\code{removed_samples}}{comma-separated list of outliers (vial labels) removed from differential analysis}
\item{\code{logFC}}{@eval logFC()}
\item{\code{logFC_se}}{@eval logFC_se()}
\item{\code{shrunk_logFC}}{log fold-change shrunk with \code{type = 'ashr'} and \code{optmethod = 'mixSQP'}, only if \code{add_shrunk_logfc = TRUE}}
\item{\code{shrunk_logFC_se}}{standard error of \code{shrunk_logFC}, only if \code{add_shrunk_logfc = TRUE}}
\item{\code{zscore}}{Wald statistic}
\item{\code{p_value}}{@eval p_value_da()}
\item{\code{comparison_average_intensity}}{@eval comparison_average_intensity()}
\item{\code{comparison_average_intensity_se}}{@eval comparison_average_intensity_se()}
\item{\code{reference_average_intensity}}{@eval reference_average_intensity()}
\item{\code{reference_average_intensity_se}}{@eval reference_average_intensity_se()}
}
}
\description{
Use DESeq2 to perform pairwise contrasts between each group of trained animals
and the sex-matched control group for a single tissue. Analysis is performed separately for males and
females.
}
\examples{
# Perform differential analysis for expressed genes in brown adipose tissue with default parameters, 
# i.e., outliers and covariates used for the manuscript; calculate both standard and shrunk log fold-changes
dea = transcript_timewise_dea("BAT")

# Same as above but don't calculate shrunk log fold-changes
dea = transcript_timewise_dea("BAT", add_shrunk_logfc = FALSE)

# Same as the first example but save the [DESeq2::DESeq2()] DESeqResults objects in an RData file 
dea = transcript_timewise_dea("BAT", rdata_outfile = "~/test/BAT_RNA_DA.RData", overwrite = TRUE)

}
