% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_differential_analysis.R
\name{transcript_training_dea_each_sex}
\alias{transcript_training_dea_each_sex}
\title{Title}
\usage{
transcript_training_dea_each_sex(
  tissue_code,
  meta,
  counts,
  covariates,
  curr_outliers,
  date,
  write = T
)
}
\arguments{
\item{tissue_code}{}

\item{meta}{}

\item{counts}{}

\item{covariates}{}

\item{curr_outliers}{}

\item{date}{}

\item{write}{}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO

}
