% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{viallabel_to_pid}
\alias{viallabel_to_pid}
\title{Map vial labels to PIDs}
\usage{
viallabel_to_pid(viallabels)
}
\arguments{
\item{viallabels}{vector, list of vial labels}
}
\value{
named list where names are vial labels and values are PIDs
}
\description{
Map sample identifiers (vial labels) to participant IDs (PIDs).
Vial labels are unique per sample; PIDs are unique per animal.
Multiple vial labels correspond to different aliquots of the same tissue sample.
}
\examples{
map = viallabel_to_pid(c("90416015402", "90416015403", "90416015302"))
map = viallabel_to_pid(c(90416015402, 90416015403, 90416015302))
map 
}
