% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_dea_args}
\alias{check_dea_args}
\title{Check arguments for DEA functions}
\usage{
check_dea_args(tissue, outfile, overwrite, outfile_is_rdata = TRUE)
}
\arguments{
\item{tissue}{@eval tissue()}

\item{outfile}{character, output file}

\item{overwrite}{bool, whether to overwrite \code{outfile} if it exists}

\item{outfile_is_rdata}{bool, whether \code{outfile} is intended to save RData}
}
\description{
Internal function used to check arguments for differential analysis functions
}
\examples{
check_dea_args("BAT", "~/Desktop/bat.rda", overwrite = FALSE) 

}
