% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_prep_data.R
\name{combine_da_results}
\alias{combine_da_results}
\title{Combine differential analysis results}
\usage{
combine_da_results(
  tissues = MotrpacRatTraining6moData::TISSUE_ABBREV,
  assays = MotrpacRatTraining6moData::ASSAY_ABBREV,
  metareg = TRUE,
  include_epigen = FALSE,
  scratchdir = "."
)
}
\arguments{
\item{tissues}{optional character vector of tissue abbreviations,
one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}.}

\item{assays}{optional character vector of assay abbreviations,
one of \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV}}

\item{metareg}{bool, whether to use the meta-regression results for METAB.
Use the upstream differential analysis results in \code{FALSE}.
\code{TRUE} by default.}

\item{include_epigen}{bool, whether to include the full ATAC or METHYL
differential analysis results from Google Cloud Storage.
Only relevant if \code{assays} includes "ATAC" or "METHYL".
\code{FALSE} by default.}

\item{scratchdir}{character, local directory in which to download data from the web.
Current working directory by default. Only relevant if \code{assays} includes "ATAC" or "METHYL".}
}
\value{
data frame. Depending on the specified assays, some of these columns may not be included:
\describe{
\item{\code{feature}}{character, unique feature identifier in the format '\link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV};\link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV};feature_ID' only for training-regulated features at 5\% IHW FDR. For redundant differential features, 'feature_ID' is prepended with the specific platform to make unique identifiers. See \link[MotrpacRatTraining6moData:REPEATED_FEATURES]{MotrpacRatTraining6moData::REPEATED_FEATURES} for details.}
\item{\code{assay}}{character, assay abbreviation, one of \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV}}
\item{\code{assay_code}}{character, assay code used in data release. See \link[MotrpacBicQC:assay_codes]{MotrpacBicQC::assay_codes}.}
\item{\code{tissue}}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}
\item{\code{tissue_code}}{character, tissue code used in data release. See \link[MotrpacBicQC:bic_animal_tissue_code]{MotrpacBicQC::bic_animal_tissue_code}.}
\item{\code{feature_ID}}{character, MoTrPAC feature identifier}
\item{\code{sex}}{character, one of 'male' or 'female'}
\item{\code{comparison_group}}{character, time point of trained animals compared to the sex-matched sedentary control animals, one of '1w', '2w', '4w', '8w'}
\item{\code{p_value}}{double, unadjusted p-value for the difference between 'comparison_group' and the group of sex-matched sedentary control animals}
\item{\code{adj_p_value}}{double, adjusted p-value from 'p_value' column. P-values are BY-adjusted across all datasets within a given assay/ome.}
\item{\code{logFC}}{double, log fold-change where the numerator is 'comparison_group', e.g., '1w', and the denominator is the group of sex-matched sedentary control animals}
\item{\code{logFC_se}}{double, standard error of the log fold-change}
\item{\code{tscore}}{double, t statistic}
\item{\code{covariates}}{character, comma-separated list of adjustment variables or NA}
\item{\code{numNAs}}{integer, number of missing values in the current 'comparison_group' and sex-matched sedentary control animals}
\item{\code{comparison_average_intensity}}{double, average intensity among the replicates in 'comparison_group'}
\item{\code{reference_average_intensity}}{double, average intensity among the replicates in the group of sex-matched sedentary control animals}
\item{\code{selection_fdr}}{double, adjusted training p-value used to select training-regulated analytes. P-values are IHW-adjusted across all datasets within a given assay with tissue as a covariate.}
\item{\code{dataset}}{character, immune panel, metabolomics platform, or ATAC-seq dataset name}
\item{\code{site}}{character, Chemical Analysis Site (CAS) name. METAB only}
\item{\code{is_targeted}}{logical, is this a targeted platform? METAB only}
\item{\code{metabolite_refmet}}{character, RefMet name of metabolite. METAB only}
\item{\code{cv}}{double, feature coefficient of variation in the dataset. METAB only}
\item{\code{metabolite}}{character, name of metabolite as appears in the CAS's data. METAB only}
\item{\code{control_cv}}{double, feature coefficient of variation in the dataset. METAB only}
\item{\code{mz}}{double, mass over charge. METAB only}
\item{\code{rt}}{double, retention time. METAB only}
\item{\code{neutral_mass}}{double, neutral mass. METAB only}
\item{\code{meta_reg_het_p}}{double, for metabolites with multiple measurements, the meta-regression heterogeneity p-value, where a smaller p-value indicates more disagreement between platforms. One value per feature (not per training group). METAB only}
\item{\code{meta_reg_pvalue}}{double, for metabolites with multiple measurements, the meta-regression p-value. One value per feature (not per training group). METAB only}
\item{\code{shrunk_logFC}}{double, log fold-change with shrinkage applied}
\item{\code{shrunk_logFC_se}}{double, standard error of the shrunken log fold-change}
\item{\code{zscore}}{double, z statistic}
\item{\code{removed_samples}}{character, comma-separated list of vial labels excluded from differential analysis or NA}
\item{\code{comparison_average_intensity_se}}{double, standard error of 'comparison_average_intensity'}
\item{\code{reference_average_intensity_se}}{double, standard error of 'comparison_average_intensity'}
\item{\code{Chr}}{integer, chromosome. METHYL only}
\item{\code{Locus}}{character, base pair range of feature. METHYL only}
\item{\code{EntrezID}}{character, Entrez ID of closest gene. METHYL only}
\item{\code{Symbol}}{character, gene symbol of closest gene. METHYL only}
}
}
\description{
Combine differential analysis results from the specified tissues and omes(s)/assay(s).
If no tissues or omes are specified, all differential analysis results are returned.
}
\examples{
# Return all non-epigenetic differential analysis results, 
# including meta-regression results for metabolomics
data = combine_da_results()

# Return all global proteomics differential analysis results
data = combine_da_results(assays="PROT")

\dontrun{
# Return METHYL and ATAC differential analysis results for gastrocnemius 
data = combine_da_results(tissues="SKM-GN", 
                          assays=c("ATAC","METHYL"),
                          include_epigen=TRUE)
}
}
