% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_datasets.R
\name{fix_covariates}
\alias{fix_covariates}
\title{Handle missing values in covariates for differential analysis}
\usage{
fix_covariates(covar, meta)
}
\arguments{
\item{covar}{string or character vector of covariate names that correspond to column names of \code{meta}}

\item{meta}{sample by variable data frame of metadata}
}
\value{
named list of two items:
\describe{
\item{\code{meta}}{data frame input \code{meta} with covariates imputed as necessary}
\item{\code{covariates}}{character vector input \code{covar} after removing covariates as necessary}
}
}
\description{
If fewer than 5\% of values are missing from a continuous covariate, replace
missing values with the mean in \code{meta}. Otherwise, remove the covariate from \code{covar}.
If values are missing from a factor covariate, remove the covariate from \code{covar}.
Note that covariates are only removed from \code{covar}, not \code{meta}.
}
\examples{
meta = data.frame(V1 = c(rnorm(19), NA),
                  V2 = c(rnorm(12), rep(NA, 8)))
covar = c("V1","V2")
result = fix_covariates(covar, meta)

}
\seealso{
\code{\link[=transcript_prep_data]{transcript_prep_data()}}
}
