% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_meta_analysis.R
\name{forest_plot}
\alias{forest_plot}
\title{Print forest plot}
\usage{
forest_plot(
  results,
  metabolite_refmet = NULL,
  sex = NULL,
  timepoint = NULL,
  name = NULL
)
}
\arguments{
\item{results}{named list returned by \code{\link[=metab_meta_analysis]{metab_meta_analysis()}}}

\item{metabolite_refmet}{character, RefMet name of metabolite}

\item{sex}{character, one of 'male' or 'female'}

\item{timepoint}{character, time point of trained animals compared to the sex-matched sedentary control animals, one of '1w', '2w', '4w', '8w'}

\item{name}{character, string in the format \code{metabolite_refmet,sex,timepoint}.
Use \emph{either} this argument \emph{or} \code{metabolite_refmet}, \code{sex}, and \code{timepoint}
to specify a result.}
}
\description{
Make a forest plot to present multiple measurements and consensus effect
size from meta-analysis.
}
\examples{
# Get meta-analysis results for gastrocnemius
res = metab_meta_analysis("SKM-GN")
# Pick a feature
res$merged_res[res$merged_res$min_nominal_p > 0.01 & 
  res$merged_res$p_value < 0.001 & 
  res$merged_res$I2 < 30,]
forest_plot(res, metabolite_refmet="Hydroxyproline", sex="female", timepoint="8w")
forest_plot(res, name="Hydroxyproline,8w,female")
}
\seealso{
\code{\link[=metab_meta_analysis]{metab_meta_analysis()}}
}
