% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_differential_analysis.R
\name{limma_res_extract_se}
\alias{limma_res_extract_se}
\title{Function to extract standard errors from limma results}
\usage{
limma_res_extract_se(limma_res, e_fit, effect_col = "logFC", t_col = "t")
}
\arguments{
\item{limma_res}{Result table as produced by limma}

\item{e_fit}{eBayes fit object as produced by limma}

\item{effect_col}{The column containing the effect size}

\item{t_col}{The column containing the t statistic}
}
\value{
A vector of standard errors for effect sizes
}
\description{
Function to extract standard errors from limma results
}
