% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_prep_data.R
\name{load_methyl_raw_data}
\alias{load_methyl_raw_data}
\title{Load raw METHYL data}
\usage{
load_methyl_raw_data(tissue, scratchdir = ".")
}
\arguments{
\item{tissue}{character, tissue abbreviation, one of "BAT", "HEART", "HIPPOC", "KIDNEY", "LIVER", "LUNG", "SKM-GN", "WAT-SC"}

\item{scratchdir}{character, local directory in which to download data from
Google Cloud Storage. Current working directory by default.}
}
\value{
named list of three items:
\describe{
\item{\code{counts}}{data frame of raw counts with features in rows and samples (unmethylated or methylated) in columns}
\item{\code{samples}}{data frame of sample-level metadata, including \code{group}, \code{lib.size}, and \code{norm.factors}}
\item{\code{genes}}{data frame of feature-level gene information, including \code{Chr} and \code{Locus}}
}
}
\description{
Load METHYL raw data for a tissue from Google Cloud Storage.
See \link{METHYL_RAW_DATA} for more details.
}
\examples{
\dontrun{
# Load raw METHYL data for gastrocnemius 
data = load_methyl_raw_data("SKM-GN", "/tmp")
}
}
