% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{prepare_gsea_input}
\alias{prepare_gsea_input}
\title{Prepare GCT file for GSEA}
\usage{
prepare_gsea_input(
  tissue = NULL,
  assay = NULL,
  gene_id_type = "gene_symbol",
  outdir = ".",
  outfile_prefix = NULL,
  input = NULL,
  cast_vars = c("sex", "comparison_group"),
  feature_to_gene_map = NULL
)
}
\arguments{
\item{tissue}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}. Must be specified if cusotm \code{input} is not provided.}

\item{assay}{character, assay abbreviation, one of "PROT", "PHOSPHO", "TRNSCRPT", "ACETYL", "UBIQ"}

\item{gene_id_type}{character, gene identifier type. Must match the gene ID type in the
gene set database you plan on using for GSEA.
One of "gene_symbol", "entrez", "ensembl", "refseq". Default: "gene_symbol"}

\item{outdir}{character, output directory for GCT file. The directory is created if it does not already exist.
Current directory by default.}

\item{outfile_prefix}{character, prefix for output GCT file. By default, this prefix
includes the specified tissue and assay and current date. Must be specified for custom input data.}

\item{input}{optional data frame if the user wants to perform this analysis for
a custom set of differential analysis results.
Required columns are "tscore", "feature_ID", and \code{cast_vars}
OR "tscore", \code{cast_vars}, and the gene identifier
indicated by \code{gene_id_type}. If a "feature_ID" column exists but not
a column corresponding to \code{gene_id_type}, then \code{feature_to_gene_map}
must map between "feature_ID" and \code{gene_id_type}.}

\item{cast_vars}{character vector of column names in the differential analysis results
that are used to convert the table from long to wide format, with t-scores as the value variable.
See \code{\link[data.table:dcast.data.table]{data.table::dcast()}} for more details. Default: "sex", "comparison_group"}

\item{feature_to_gene_map}{data frame, map between "feature_ID" and \code{gene_id_type}.
\link[MotrpacRatTraining6moData:FEATURE_TO_GENE_FILT]{MotrpacRatTraining6moData::FEATURE_TO_GENE_FILT} if not otherwise specified.}
}
\value{
character, path of the GCT file
}
\description{
Take a set of differential analysis results, either for a specific tissue and ome or
from a user-supplied table, convert to GCT format, and write to file for GSEA.
}
\details{
T-scores from the timewise differential analysis results are used for scores.
Feature-level data is summarized into gene-level data using the maximum absolute t-score.
}
\examples{
# Input for GSEA where gene set databases use gene symbols as IDs.
# This applies for the default gene set database 
# when 'method="gsea"' in 'ssGSEA2_wrapper()'.
prepare_gsea_input("HEART","PROT",outdir="/tmp")
prepare_gsea_input("LIVER","PHOSPHO",outdir="/tmp")
prepare_gsea_input("LIVER","ACETYL",outdir="/tmp")
prepare_gsea_input("LIVER","UBIQ",outdir="/tmp")
prepare_gsea_input("LIVER","TRNSCRPT",outdir="/tmp")

# Input for GSEA with Mitocarta (i.e., 'method="gsea_mitocarta"' 
# in 'ssGSEA2_wrapper()'), which uses RefSeq IDs 
prepare_gsea_input("LIVER","TRNSCRPT",outdir="/tmp",gene_id_type="refseq")

# "Custom" input
res = combine_da_results(tissues = "KIDNEY", assays = "PROT")
# add dummy column
res$gene_symbol = res$feature_ID
prepare_gsea_input(input=res, outdir="/tmp", outfile_prefix="KIDNEY_PROT")


}
