% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_differential_analysis.R
\name{run_deseq}
\alias{run_deseq}
\title{Wrapper for \code{\link[DESeq2:DESeq]{DESeq2::DESeq()}}}
\usage{
run_deseq(
  counts,
  meta,
  covar,
  outcome_of_interest,
  contrasts,
  shrink = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{counts}{data frame of raw filtered RNA-seq counts. Row names are gene IDs and
column names are sample IDs. Column names must correspond to values in \code{meta$viallabel}}

\item{meta}{data frame of metadata with columns \code{\param{outcome_of_interest}, \param{covar}, 'viallabel'} at a minimum.
\param{counts} are subset to \code{meta$viallabel}}

\item{covar}{adjustment variables to include in the DESeq model}

\item{outcome_of_interest}{outcome of interest to include in the model;
levels provided in \param{contrasts}}

\item{shrink}{bool, whether to apply \code{lfcShrink()}}

\item{verbose}{bool, whether to print the design string}

\item{constrasts}{list of vectors, where each vector is in the form
\code(c(outcome_of_interest, numerator, denominator)), e.g. \code{c('sex_group','female.1w','female.control')}}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO

}
