% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_outliers.R
\name{transcript_call_outliers}
\alias{transcript_call_outliers}
\title{Call RNA-seq outliers}
\usage{
transcript_call_outliers(tissues)
}
\arguments{
\item{tissues}{list of tissue abbreviations for which to call RNA-seq outliers.
See \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV} for accepted values.}
}
\value{
TODO
}
\description{
TODO
}
\examples{
transcript_call_outliers("SKM-GN")
transcript_call_outliers(c("SKM-GN","BLOOD"))

}
\seealso{
\code{\link[=call_pca_outliers]{call_pca_outliers()}} for workhorse function, \code{\link[=plot_pcs]{plot_pcs()}} for plotting function,
\link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV} for list of accepted tissue abbrevations
}
