% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_differential_analysis.R
\name{transcript_prep_data}
\alias{transcript_prep_data}
\title{Title}
\usage{
transcript_prep_data(
  tissue_code,
  meta,
  counts,
  tmm,
  covariates = c("pct_globin", "rin", "pct_umi_dup", "median_5_3_bias"),
  outliers = NULL,
  gsutil_path = "~/google-cloud-sdk/bin/gsutil",
  parallel = F
)
}
\arguments{
\item{tissue_code}{}

\item{meta}{}

\item{counts}{}

\item{tmm}{}

\item{covariates}{}

\item{outliers}{}

\item{gsutil_path}{}

\item{parallel}{}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO

}
