% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_prep_data.R
\name{transcript_prep_data}
\alias{transcript_prep_data}
\title{Preprocess RNA-seq data}
\usage{
transcript_prep_data(
  tissue,
  sex = NULL,
  covariates = c("pct_globin", "RIN", "pct_umi_dup", "median_5_3_bias"),
  outliers = na.omit(MotrpacRatTraining6moData::OUTLIERS$viallabel),
  adjust_covariates = TRUE,
  center_scale = FALSE
)
}
\arguments{
\item{tissue}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}

\item{sex}{character, one of 'male' or 'female'}

\item{covariates}{character vector of covariates that correspond to column names of \link[MotrpacRatTraining6moData:TRNSCRPT_META]{MotrpacRatTraining6moData::TRNSCRPT_META}.
Defaults to covariates that were used for the manuscript.}

\item{outliers}{vector of viallabels to exclude from the returned data. Defaults
to \code{[MotrpacRatTraining6moData::OUTLIERS]$viallabel}}

\item{adjust_covariates}{boolean, whether to adjust covariates using \code{\link[=fix_covariates]{fix_covariates()}}.
Only applies if \code{covariates} is not NULL.}

\item{center_scale}{boolean, whether to center and scale continuous covariates within \code{\link[=fix_covariates]{fix_covariates()}}.
Only applies if \code{adjust_covariates} is also TRUE.}
}
\value{
named list of five items:
\describe{
\item{\code{metadata}}{data frame of combined \link[MotrpacRatTraining6moData:PHENO]{MotrpacRatTraining6moData::PHENO} and \link[MotrpacRatTraining6moData:TRNSCRPT_META]{MotrpacRatTraining6moData::TRNSCRPT_META}, filtered to samples in \code{tissue}.
If \code{adjust_covariates = TRUE}, missing values in \code{covariates} are imputed.
If also \code{center_scale = TRUE}, continuous variables named by \code{covariates} are centered and scaled.}
\item{\code{covariates}}{character vector of covariates to adjust for during differential analysis. For all tissues except VENACV,
this vector is a (sub)set of the input list of covariates. Covariates are removed from this vector if there are too
many missing values or if all values are constant. See \code{\link[=fix_covariates]{fix_covariates()}} for more details.
If \code{tissue = "VENACV"}, the Ensembl ID for Ucp1 is also added as a covariate.}
\item{\code{counts}}{data frame of raw counts with Ensembl IDs (which are also TRNSCRPT \code{feature_ID}s)
as row names and vial labels as column names. See \link[MotrpacRatTraining6moData:TRNSCRPT_RAW_COUNTS]{MotrpacRatTraining6moData::TRNSCRPT_RAW_COUNTS} for details.}
\item{\code{norm_data}}{data frame of TMM-normalized data with Ensembl IDs (which are also TRNSCRPT
\code{feature_ID}s) as row names and vial labels as column names. See \link[MotrpacRatTraining6moData:TRNSCRPT_NORM_DATA]{MotrpacRatTraining6moData::TRNSCRPT_NORM_DATA} for details.}
\item{\code{outliers}}{subset of \code{outliers} in input removed from the data}
}
}
\description{
Collect filtered raw counts, normalized sample-level data, phenotypic data, RNA-seq metadata,
covariates, and outliers associated with a given tissue.
}
\examples{
# Process gastrocnemius RNA-seq data with default parameters, i.e., return data from both 
# sexes, remove established outliers, impute missing values in default covariates 
gastroc_data1 = transcript_prep_data("SKM-GN")

# Same as above but do not remove outliers if they exist 
gastroc_data2 = transcript_prep_data("SKM-GN", outliers = NULL)

# Same as above but do not adjust existing variables in the metadata  
gastroc_data3 = transcript_prep_data("SKM-GN", covariates = NULL, outliers = NULL)

# Same as above but only return data from male samples
gastroc_data4 = transcript_prep_data("SKM-GN", covariates = NULL, outliers = NULL, sex = "male")

# Same as gastroc_data2 but also center and scale default continuous covariates in the returned metadata,
# which is also done within [run_deseq()] (called by [transcript_timewise_dea()]) 
gastroc_data4 = transcript_prep_data("SKM-GN", outliers = NULL, center_scale = TRUE)

}
\seealso{
\link[MotrpacRatTraining6moData:OUTLIERS]{MotrpacRatTraining6moData::OUTLIERS},
\code{\link[=fix_covariates]{fix_covariates()}},
\link[MotrpacRatTraining6moData:PHENO]{MotrpacRatTraining6moData::PHENO},
\link[MotrpacRatTraining6moData:TRNSCRPT_META]{MotrpacRatTraining6moData::TRNSCRPT_META},
\link[MotrpacRatTraining6moData:TRNSCRPT_RAW_COUNTS]{MotrpacRatTraining6moData::TRNSCRPT_RAW_COUNTS},
\link[MotrpacRatTraining6moData:TRNSCRPT_NORM_DATA]{MotrpacRatTraining6moData::TRNSCRPT_NORM_DATA}
}
