% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_differential_analysis.R
\name{transcript_timewise_dea}
\alias{transcript_timewise_dea}
\title{TODO
RNA-seq timewise differential analysis}
\usage{
transcript_timewise_dea(
  tissue,
  covariates,
  outliers,
  date = NULL,
  save_rdata = T,
  write = T
)
}
\arguments{
\item{covariates}{}

\item{outliers}{}

\item{date}{}

\item{save_rdata}{}

\item{write}{}

\item{tissue_code}{}
}
\value{
TODO
}
\description{
TODO
RNA-seq timewise differential analysis
}
\examples{
TODO

}
