% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_differential_analysis.R
\name{transcript_timewise_dea}
\alias{transcript_timewise_dea}
\title{RNA-seq timewise differential analysis}
\usage{
transcript_timewise_dea(
  tissue,
  covariates = c("pct_globin", "RIN", "pct_umi_dup", "median_5_3_bias"),
 
    outliers = MotrpacRatTraining6moData::OUTLIERS$viallabel[MotrpacRatTraining6moData::OUTLIERS$assay
    == "TRNSCRPT"],
  add_shrunk_logfc = TRUE,
  rdata_outfile = NULL,
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{tissue}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}

\item{covariates}{character vector of covariates that correspond to column names of \link[MotrpacRatTraining6moData:TRNSCRPT_META]{MotrpacRatTraining6moData::TRNSCRPT_META}.
Defaults to covariates that were used for the manuscript.}

\item{outliers}{vector of viallabels to exclude during differential analysis. Defaults
to \code{MotrpacRatTraining6moData::OUTLIERS$viallabel[MotrpacRatTraining6moData::OUTLIERS$assay == "TRNSCRPT"]}}

\item{add_shrunk_logfc}{boolean, whether to calculate shrunk log fold-changes in addition to standard log fold-changes}

\item{rdata_outfile}{NULL or path in which to save DESeq2 objects in an RData file}

\item{overwrite}{boolean, whether to overwrite the file if \code{rdata_outfile} exists}

\item{verbose}{boolean, whether to print the DESeq2 design string}
}
\value{
a data frame with one row per gene per contrast (usually 8 rows per gene):
\describe{
\item{\code{feature}}{unique gene identifier in the format \code{[ASSAY_ABBREV];[TISSUE_ABBREV];[feature_ID]}}
\item{\code{feature_ID}}{Ensembl gene ID}
\item{\code{sex}}{one of "male" or "female"}
\item{\code{comparison_group}}{time point compared to the sex-matched
sedentary control animals, one of "1w", "2w", "4w", "8w"}
\item{\code{assay}}{assay abbreviation, one of \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV}}
\item{\code{assay_code}}{MoTrPAC assay or "ome" code. "transcript-rna-seq" for RNA-seq datasets.}
\item{\code{tissue}}{tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}
\item{\code{tissue_code}}{MoTrPAC tissue release code. See \link[MotrpacBicQC:bic_animal_tissue_code]{MotrpacBicQC::bic_animal_tissue_code} for details.}
\item{\code{covariates}}{comma-separated list of adjustment variables}
\item{\code{removed_samples}}{comma-separated list of outliers (vial labels) removed from differential analysis}
\item{\code{logFC}}{log fold-change of the training group specified by \code{sex} and \code{comparison_group} (e.g., 1-week females)
relative to the sex-matched sedentary controls}
\item{\code{logFC_se}}{standard error of \code{logFC}}
\item{\code{shrunk_logFC}}{log fold-change shrunk with \code{type = 'ashr'} and \code{optmethod = 'mixSQP'}, only if \code{add_shrunk_logfc = TRUE}}
\item{\code{shrunk_logFC_se}}{standard error of \code{shrunk_logFC}, only if \code{add_shrunk_logfc = TRUE}}
\item{\code{zscore}}{Wald statistic}
\item{\code{p_value}}{nominal p-value corresponding to the contrast between the training group
(e.g., 1-week females) and the sex-matched sedentary controls}
\item{\code{comparison_average_intensity}}{average normalized RNA-seq counts for samples in the training group (e.g., 1-week females)}
\item{\code{comparison_average_intensity_se}}{standard error of \code{comparison_average_intensity}}
\item{\code{reference_average_intensity}}{average normalized RNA-seq counts for sex-matched sedentary control samples}
\item{\code{reference_average_intensity_se}}{standard error of \code{reference_average_intensity}}
}
}
\description{
Use DESeq2 to perform pairwise contrasts between each group of trained animals
and the sex-matched control group for a single tissue. Analysis is performed separately for males and
females.
}
\examples{
# Perform differential analysis for expressed genes in brown adipose tissue with default parameters, 
# i.e., outliers and covariates used for the manuscript; calculate both standard and shrunk log fold-changes
dea = transcript_timewise_dea("BAT")

# Same as above but don't calculate shrunk log fold-changes
dea = transcript_timewise_dea("BAT", add_shrunk_logfc = FALSE)

# Same as the first example but save the [DESeq2::DESeq2()] DESeqResults objects in an RData file 
dea = transcript_timewise_dea("BAT", rdata_outfile = "~/test/BAT_TRNSCRIPT_DA.RData", overwrite = TRUE)

}
