% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_differential_analysis.R
\name{transcript_training_dea}
\alias{transcript_training_dea}
\title{RNA-seq training differential analysis}
\usage{
transcript_training_dea(
  tissue,
  covariates = c("pct_globin", "RIN", "pct_umi_dup", "median_5_3_bias"),
 
    outliers = na.omit(MotrpacRatTraining6moData::OUTLIERS$viallabel[MotrpacRatTraining6moData::OUTLIERS$assay
    == "TRNSCRPT"]),
  rdata_outfile = NULL,
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{tissue}{character, tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}

\item{covariates}{character vector of covariates that correspond to column names of \link[MotrpacRatTraining6moData:TRNSCRPT_META]{MotrpacRatTraining6moData::TRNSCRPT_META}.
Defaults to covariates that were used for the manuscript.}

\item{outliers}{vector of viallabels to exclude during differential analysis. Defaults
to \code{[MotrpacRatTraining6moData::OUTLIERS]$viallabel[[MotrpacRatTraining6moData::OUTLIERS]$assay == "TRNSCRPT"]}}

\item{rdata_outfile}{NULL or path in which to save DESeq2 objects in an RData file}

\item{overwrite}{boolean, whether to overwrite the file if \code{rdata_outfile} exists}

\item{verbose}{boolean, whether to print the DESeq2 design string}
}
\value{
a data frame with one row per gene:
\describe{
\item{\code{feature}}{unique gene identifier in the format \code{[ASSAY_ABBREV];[TISSUE_ABBREV];[feature_ID]}}
\item{\code{feature_ID}}{Ensembl gene ID}
\item{\code{sex}}{one of "male" or "female"}
\item{\code{assay}}{assay abbreviation, one of \link[MotrpacRatTraining6moData:ASSAY_ABBREV]{MotrpacRatTraining6moData::ASSAY_ABBREV}}
\item{\code{assay_code}}{MoTrPAC assay or "ome" code. "transcript-rna-seq" for RNA-seq datasets.}
\item{\code{tissue}}{tissue abbreviation, one of \link[MotrpacRatTraining6moData:TISSUE_ABBREV]{MotrpacRatTraining6moData::TISSUE_ABBREV}}
\item{\code{tissue_code}}{MoTrPAC tissue release code. See \link[MotrpacBicQC:bic_animal_tissue_code]{MotrpacBicQC::bic_animal_tissue_code} for details.}
\item{\code{removed_samples_male}}{comma-separated list of male outliers (vial labels) removed from differential analysis}
\item{\code{removed_samples_female}}{comma-separated list of female outliers (vial labels) removed from differential analysis}
\item{\code{lrt_male}}{likelihood ratio test statistic for males}
\item{\code{lrt_female}}{likelihood ratio test statistic for females}
\item{\code{p_value_male}}{nominal LRT p-value for males}
\item{\code{p_value_female}}{nominal LRT p-value for females}
\item{\code{full_model_male}}{full model used in LRT for males}
\item{\code{full_model_female}}{full model used in LRT for females}
\item{\code{reduced_model_male}}{reduced model used in LRT for males}
\item{\code{reduced_model_female}}{reduced model used in LRT for females}
\item{\code{p_value}}{combined male and female nominal p-value using the sum of logs}
}
}
\description{
Use DESeq2 to perform a likelihood ratio test to test the effect of training
across time points. Analysis is performed separately for males and females.
}
\examples{
# Perform differential analysis for expressed genes in brown adipose tissue with default parameters, 
# i.e., outliers and covariates used for the manuscript
dea = transcript_training_dea("BAT")

# Same as above but save the [DESeq2::DESeq2()] DESeqResults objects in an RData file 
dea = transcript_training_dea("BAT", rdata_outfile = "~/test/BAT_RNA_training-dea.RData", overwrite = TRUE)

}
